Interpolate geodesically between bases.

This function takes a set of bases and produces a tour by geodesically interpolating between each basis

interpolate(basis_set, angle = 0.05, cycle = FALSE)

Arguments

basis_set
input basis set

angle
target distance (in radians) between bases

cycle
For planned_tour cycle through continuously (TRUE) or stop after first pass (FALSE)

Examples

t1 <- save_history(flea[, 1:6], grand_tour(1), max = 10)

#> Converting input data to the required matrix format.

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.412588 
#> target_dist - cur_dist: 1.412588 
#> generation:  dist =   1.33445 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.299216 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.348672 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.344725 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.370737 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.219314 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.290812 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.394297 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.011288 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.291949 

dim(t1)

#> [1]  6  1 10

dim(interpolate(t1, 0.01))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.299216 
#> target_dist - cur_dist: 1.299216 
#> target_dist - cur_dist: 1.289216 
#> target_dist - cur_dist: 1.279216 
#> target_dist - cur_dist: 1.269216 
#> target_dist - cur_dist: 1.259216 
#> target_dist - cur_dist: 1.249216 
#> target_dist - cur_dist: 1.239216 
#> target_dist - cur_dist: 1.229216 
#> target_dist - cur_dist: 1.219216 
#> target_dist - cur_dist: 1.209216 
#> target_dist - cur_dist: 1.199216 
#> target_dist - cur_dist: 1.189216 
#> target_dist - cur_dist: 1.179216 
#> target_dist - cur_dist: 1.169216 
#> target_dist - cur_dist: 1.159216 
#> target_dist - cur_dist: 1.149216 
#> target_dist - cur_dist: 1.139216 
#> target_dist - cur_dist: 1.129216 
#> target_dist - cur_dist: 1.119216 
#> target_dist - cur_dist: 1.109216 
#> target_dist - cur_dist: 1.099216 
#> target_dist - cur_dist: 1.089216 
#> target_dist - cur_dist: 1.079216 
#> target_dist - cur_dist: 1.069216 
#> target_dist - cur_dist: 1.059216 
#> target_dist - cur_dist: 1.049216 
#> target_dist - cur_dist: 1.039216 
#> target_dist - cur_dist: 1.029216 
#> target_dist - cur_dist: 1.019216 
#> target_dist - cur_dist: 1.009216 
#> target_dist - cur_dist: 0.9992157 
#> target_dist - cur_dist: 0.9892157 
#> target_dist - cur_dist: 0.9792157 
#> target_dist - cur_dist: 0.9692157 
#> target_dist - cur_dist: 0.9592157 
#> target_dist - cur_dist: 0.9492157 
#> target_dist - cur_dist: 0.9392157 
#> target_dist - cur_dist: 0.9292157 
#> target_dist - cur_dist: 0.9192157 
#> target_dist - cur_dist: 0.9092157 
#> target_dist - cur_dist: 0.8992157 
#> target_dist - cur_dist: 0.8892157 
#> target_dist - cur_dist: 0.8792157 
#> target_dist - cur_dist: 0.8692157 
#> target_dist - cur_dist: 0.8592157 
#> target_dist - cur_dist: 0.8492157 
#> target_dist - cur_dist: 0.8392157 
#> target_dist - cur_dist: 0.8292157 
#> target_dist - cur_dist: 0.8192157 
#> target_dist - cur_dist: 0.8092157 
#> target_dist - cur_dist: 0.7992157 
#> target_dist - cur_dist: 0.7892157 
#> target_dist - cur_dist: 0.7792157 
#> target_dist - cur_dist: 0.7692157 
#> target_dist - cur_dist: 0.7592157 
#> target_dist - cur_dist: 0.7492157 
#> target_dist - cur_dist: 0.7392157 
#> target_dist - cur_dist: 0.7292157 
#> target_dist - cur_dist: 0.7192157 
#> target_dist - cur_dist: 0.7092157 
#> target_dist - cur_dist: 0.6992157 
#> target_dist - cur_dist: 0.6892157 
#> target_dist - cur_dist: 0.6792157 
#> target_dist - cur_dist: 0.6692157 
#> target_dist - cur_dist: 0.6592157 
#> target_dist - cur_dist: 0.6492157 
#> target_dist - cur_dist: 0.6392157 
#> target_dist - cur_dist: 0.6292157 
#> target_dist - cur_dist: 0.6192157 
#> target_dist - cur_dist: 0.6092157 
#> target_dist - cur_dist: 0.5992157 
#> target_dist - cur_dist: 0.5892157 
#> target_dist - cur_dist: 0.5792157 
#> target_dist - cur_dist: 0.5692157 
#> target_dist - cur_dist: 0.5592157 
#> target_dist - cur_dist: 0.5492157 
#> target_dist - cur_dist: 0.5392157 
#> target_dist - cur_dist: 0.5292157 
#> target_dist - cur_dist: 0.5192157 
#> target_dist - cur_dist: 0.5092157 
#> target_dist - cur_dist: 0.4992157 
#> target_dist - cur_dist: 0.4892157 
#> target_dist - cur_dist: 0.4792157 
#> target_dist - cur_dist: 0.4692157 
#> target_dist - cur_dist: 0.4592157 
#> target_dist - cur_dist: 0.4492157 
#> target_dist - cur_dist: 0.4392157 
#> target_dist - cur_dist: 0.4292157 
#> target_dist - cur_dist: 0.4192157 
#> target_dist - cur_dist: 0.4092157 
#> target_dist - cur_dist: 0.3992157 
#> target_dist - cur_dist: 0.3892157 
#> target_dist - cur_dist: 0.3792157 
#> target_dist - cur_dist: 0.3692157 
#> target_dist - cur_dist: 0.3592157 
#> target_dist - cur_dist: 0.3492157 
#> target_dist - cur_dist: 0.3392157 
#> target_dist - cur_dist: 0.3292157 
#> target_dist - cur_dist: 0.3192157 
#> target_dist - cur_dist: 0.3092157 
#> target_dist - cur_dist: 0.2992157 
#> target_dist - cur_dist: 0.2892157 
#> target_dist - cur_dist: 0.2792157 
#> target_dist - cur_dist: 0.2692157 
#> target_dist - cur_dist: 0.2592157 
#> target_dist - cur_dist: 0.2492157 
#> target_dist - cur_dist: 0.2392157 
#> target_dist - cur_dist: 0.2292157 
#> target_dist - cur_dist: 0.2192157 
#> target_dist - cur_dist: 0.2092157 
#> target_dist - cur_dist: 0.1992157 
#> target_dist - cur_dist: 0.1892157 
#> target_dist - cur_dist: 0.1792157 
#> target_dist - cur_dist: 0.1692157 
#> target_dist - cur_dist: 0.1592157 
#> target_dist - cur_dist: 0.1492157 
#> target_dist - cur_dist: 0.1392157 
#> target_dist - cur_dist: 0.1292157 
#> target_dist - cur_dist: 0.1192157 
#> target_dist - cur_dist: 0.1092157 
#> target_dist - cur_dist: 0.09921574 
#> target_dist - cur_dist: 0.08921574 
#> target_dist - cur_dist: 0.07921574 
#> target_dist - cur_dist: 0.06921574 
#> target_dist - cur_dist: 0.05921574 
#> target_dist - cur_dist: 0.04921574 
#> target_dist - cur_dist: 0.03921574 
#> target_dist - cur_dist: 0.02921574 
#> target_dist - cur_dist: 0.01921574 
#> target_dist - cur_dist: 0.009215743 
#> generation:  dist =   1.348672 
#> target_dist - cur_dist: 1.348672 
#> target_dist - cur_dist: 1.338672 
#> target_dist - cur_dist: 1.328672 
#> target_dist - cur_dist: 1.318672 
#> target_dist - cur_dist: 1.308672 
#> target_dist - cur_dist: 1.298672 
#> target_dist - cur_dist: 1.288672 
#> target_dist - cur_dist: 1.278672 
#> target_dist - cur_dist: 1.268672 
#> target_dist - cur_dist: 1.258672 
#> target_dist - cur_dist: 1.248672 
#> target_dist - cur_dist: 1.238672 
#> target_dist - cur_dist: 1.228672 
#> target_dist - cur_dist: 1.218672 
#> target_dist - cur_dist: 1.208672 
#> target_dist - cur_dist: 1.198672 
#> target_dist - cur_dist: 1.188672 
#> target_dist - cur_dist: 1.178672 
#> target_dist - cur_dist: 1.168672 
#> target_dist - cur_dist: 1.158672 
#> target_dist - cur_dist: 1.148672 
#> target_dist - cur_dist: 1.138672 
#> target_dist - cur_dist: 1.128672 
#> target_dist - cur_dist: 1.118672 
#> target_dist - cur_dist: 1.108672 
#> target_dist - cur_dist: 1.098672 
#> target_dist - cur_dist: 1.088672 
#> target_dist - cur_dist: 1.078672 
#> target_dist - cur_dist: 1.068672 
#> target_dist - cur_dist: 1.058672 
#> target_dist - cur_dist: 1.048672 
#> target_dist - cur_dist: 1.038672 
#> target_dist - cur_dist: 1.028672 
#> target_dist - cur_dist: 1.018672 
#> target_dist - cur_dist: 1.008672 
#> target_dist - cur_dist: 0.9986724 
#> target_dist - cur_dist: 0.9886724 
#> target_dist - cur_dist: 0.9786724 
#> target_dist - cur_dist: 0.9686724 
#> target_dist - cur_dist: 0.9586724 
#> target_dist - cur_dist: 0.9486724 
#> target_dist - cur_dist: 0.9386724 
#> target_dist - cur_dist: 0.9286724 
#> target_dist - cur_dist: 0.9186724 
#> target_dist - cur_dist: 0.9086724 
#> target_dist - cur_dist: 0.8986724 
#> target_dist - cur_dist: 0.8886724 
#> target_dist - cur_dist: 0.8786724 
#> target_dist - cur_dist: 0.8686724 
#> target_dist - cur_dist: 0.8586724 
#> target_dist - cur_dist: 0.8486724 
#> target_dist - cur_dist: 0.8386724 
#> target_dist - cur_dist: 0.8286724 
#> target_dist - cur_dist: 0.8186724 
#> target_dist - cur_dist: 0.8086724 
#> target_dist - cur_dist: 0.7986724 
#> target_dist - cur_dist: 0.7886724 
#> target_dist - cur_dist: 0.7786724 
#> target_dist - cur_dist: 0.7686724 
#> target_dist - cur_dist: 0.7586724 
#> target_dist - cur_dist: 0.7486724 
#> target_dist - cur_dist: 0.7386724 
#> target_dist - cur_dist: 0.7286724 
#> target_dist - cur_dist: 0.7186724 
#> target_dist - cur_dist: 0.7086724 
#> target_dist - cur_dist: 0.6986724 
#> target_dist - cur_dist: 0.6886724 
#> target_dist - cur_dist: 0.6786724 
#> target_dist - cur_dist: 0.6686724 
#> target_dist - cur_dist: 0.6586724 
#> target_dist - cur_dist: 0.6486724 
#> target_dist - cur_dist: 0.6386724 
#> target_dist - cur_dist: 0.6286724 
#> target_dist - cur_dist: 0.6186724 
#> target_dist - cur_dist: 0.6086724 
#> target_dist - cur_dist: 0.5986724 
#> target_dist - cur_dist: 0.5886724 
#> target_dist - cur_dist: 0.5786724 
#> target_dist - cur_dist: 0.5686724 
#> target_dist - cur_dist: 0.5586724 
#> target_dist - cur_dist: 0.5486724 
#> target_dist - cur_dist: 0.5386724 
#> target_dist - cur_dist: 0.5286724 
#> target_dist - cur_dist: 0.5186724 
#> target_dist - cur_dist: 0.5086724 
#> target_dist - cur_dist: 0.4986724 
#> target_dist - cur_dist: 0.4886724 
#> target_dist - cur_dist: 0.4786724 
#> target_dist - cur_dist: 0.4686724 
#> target_dist - cur_dist: 0.4586724 
#> target_dist - cur_dist: 0.4486724 
#> target_dist - cur_dist: 0.4386724 
#> target_dist - cur_dist: 0.4286724 
#> target_dist - cur_dist: 0.4186724 
#> target_dist - cur_dist: 0.4086724 
#> target_dist - cur_dist: 0.3986724 
#> target_dist - cur_dist: 0.3886724 
#> target_dist - cur_dist: 0.3786724 
#> target_dist - cur_dist: 0.3686724 
#> target_dist - cur_dist: 0.3586724 
#> target_dist - cur_dist: 0.3486724 
#> target_dist - cur_dist: 0.3386724 
#> target_dist - cur_dist: 0.3286724 
#> target_dist - cur_dist: 0.3186724 
#> target_dist - cur_dist: 0.3086724 
#> target_dist - cur_dist: 0.2986724 
#> target_dist - cur_dist: 0.2886724 
#> target_dist - cur_dist: 0.2786724 
#> target_dist - cur_dist: 0.2686724 
#> target_dist - cur_dist: 0.2586724 
#> target_dist - cur_dist: 0.2486724 
#> target_dist - cur_dist: 0.2386724 
#> target_dist - cur_dist: 0.2286724 
#> target_dist - cur_dist: 0.2186724 
#> target_dist - cur_dist: 0.2086724 
#> target_dist - cur_dist: 0.1986724 
#> target_dist - cur_dist: 0.1886724 
#> target_dist - cur_dist: 0.1786724 
#> target_dist - cur_dist: 0.1686724 
#> target_dist - cur_dist: 0.1586724 
#> target_dist - cur_dist: 0.1486724 
#> target_dist - cur_dist: 0.1386724 
#> target_dist - cur_dist: 0.1286724 
#> target_dist - cur_dist: 0.1186724 
#> target_dist - cur_dist: 0.1086724 
#> target_dist - cur_dist: 0.09867236 
#> target_dist - cur_dist: 0.08867236 
#> target_dist - cur_dist: 0.07867236 
#> target_dist - cur_dist: 0.06867236 
#> target_dist - cur_dist: 0.05867236 
#> target_dist - cur_dist: 0.04867236 
#> target_dist - cur_dist: 0.03867236 
#> target_dist - cur_dist: 0.02867236 
#> target_dist - cur_dist: 0.01867236 
#> target_dist - cur_dist: 0.008672364 
#> generation:  dist =   1.344725 
#> target_dist - cur_dist: 1.344725 
#> target_dist - cur_dist: 1.334725 
#> target_dist - cur_dist: 1.324725 
#> target_dist - cur_dist: 1.314725 
#> target_dist - cur_dist: 1.304725 
#> target_dist - cur_dist: 1.294725 
#> target_dist - cur_dist: 1.284725 
#> target_dist - cur_dist: 1.274725 
#> target_dist - cur_dist: 1.264725 
#> target_dist - cur_dist: 1.254725 
#> target_dist - cur_dist: 1.244725 
#> target_dist - cur_dist: 1.234725 
#> target_dist - cur_dist: 1.224725 
#> target_dist - cur_dist: 1.214725 
#> target_dist - cur_dist: 1.204725 
#> target_dist - cur_dist: 1.194725 
#> target_dist - cur_dist: 1.184725 
#> target_dist - cur_dist: 1.174725 
#> target_dist - cur_dist: 1.164725 
#> target_dist - cur_dist: 1.154725 
#> target_dist - cur_dist: 1.144725 
#> target_dist - cur_dist: 1.134725 
#> target_dist - cur_dist: 1.124725 
#> target_dist - cur_dist: 1.114725 
#> target_dist - cur_dist: 1.104725 
#> target_dist - cur_dist: 1.094725 
#> target_dist - cur_dist: 1.084725 
#> target_dist - cur_dist: 1.074725 
#> target_dist - cur_dist: 1.064725 
#> target_dist - cur_dist: 1.054725 
#> target_dist - cur_dist: 1.044725 
#> target_dist - cur_dist: 1.034725 
#> target_dist - cur_dist: 1.024725 
#> target_dist - cur_dist: 1.014725 
#> target_dist - cur_dist: 1.004725 
#> target_dist - cur_dist: 0.9947249 
#> target_dist - cur_dist: 0.9847249 
#> target_dist - cur_dist: 0.9747249 
#> target_dist - cur_dist: 0.9647249 
#> target_dist - cur_dist: 0.9547249 
#> target_dist - cur_dist: 0.9447249 
#> target_dist - cur_dist: 0.9347249 
#> target_dist - cur_dist: 0.9247249 
#> target_dist - cur_dist: 0.9147249 
#> target_dist - cur_dist: 0.9047249 
#> target_dist - cur_dist: 0.8947249 
#> target_dist - cur_dist: 0.8847249 
#> target_dist - cur_dist: 0.8747249 
#> target_dist - cur_dist: 0.8647249 
#> target_dist - cur_dist: 0.8547249 
#> target_dist - cur_dist: 0.8447249 
#> target_dist - cur_dist: 0.8347249 
#> target_dist - cur_dist: 0.8247249 
#> target_dist - cur_dist: 0.8147249 
#> target_dist - cur_dist: 0.8047249 
#> target_dist - cur_dist: 0.7947249 
#> target_dist - cur_dist: 0.7847249 
#> target_dist - cur_dist: 0.7747249 
#> target_dist - cur_dist: 0.7647249 
#> target_dist - cur_dist: 0.7547249 
#> target_dist - cur_dist: 0.7447249 
#> target_dist - cur_dist: 0.7347249 
#> target_dist - cur_dist: 0.7247249 
#> target_dist - cur_dist: 0.7147249 
#> target_dist - cur_dist: 0.7047249 
#> target_dist - cur_dist: 0.6947249 
#> target_dist - cur_dist: 0.6847249 
#> target_dist - cur_dist: 0.6747249 
#> target_dist - cur_dist: 0.6647249 
#> target_dist - cur_dist: 0.6547249 
#> target_dist - cur_dist: 0.6447249 
#> target_dist - cur_dist: 0.6347249 
#> target_dist - cur_dist: 0.6247249 
#> target_dist - cur_dist: 0.6147249 
#> target_dist - cur_dist: 0.6047249 
#> target_dist - cur_dist: 0.5947249 
#> target_dist - cur_dist: 0.5847249 
#> target_dist - cur_dist: 0.5747249 
#> target_dist - cur_dist: 0.5647249 
#> target_dist - cur_dist: 0.5547249 
#> target_dist - cur_dist: 0.5447249 
#> target_dist - cur_dist: 0.5347249 
#> target_dist - cur_dist: 0.5247249 
#> target_dist - cur_dist: 0.5147249 
#> target_dist - cur_dist: 0.5047249 
#> target_dist - cur_dist: 0.4947249 
#> target_dist - cur_dist: 0.4847249 
#> target_dist - cur_dist: 0.4747249 
#> target_dist - cur_dist: 0.4647249 
#> target_dist - cur_dist: 0.4547249 
#> target_dist - cur_dist: 0.4447249 
#> target_dist - cur_dist: 0.4347249 
#> target_dist - cur_dist: 0.4247249 
#> target_dist - cur_dist: 0.4147249 
#> target_dist - cur_dist: 0.4047249 
#> target_dist - cur_dist: 0.3947249 
#> target_dist - cur_dist: 0.3847249 
#> target_dist - cur_dist: 0.3747249 
#> target_dist - cur_dist: 0.3647249 
#> target_dist - cur_dist: 0.3547249 
#> target_dist - cur_dist: 0.3447249 
#> target_dist - cur_dist: 0.3347249 
#> target_dist - cur_dist: 0.3247249 
#> target_dist - cur_dist: 0.3147249 
#> target_dist - cur_dist: 0.3047249 
#> target_dist - cur_dist: 0.2947249 
#> target_dist - cur_dist: 0.2847249 
#> target_dist - cur_dist: 0.2747249 
#> target_dist - cur_dist: 0.2647249 
#> target_dist - cur_dist: 0.2547249 
#> target_dist - cur_dist: 0.2447249 
#> target_dist - cur_dist: 0.2347249 
#> target_dist - cur_dist: 0.2247249 
#> target_dist - cur_dist: 0.2147249 
#> target_dist - cur_dist: 0.2047249 
#> target_dist - cur_dist: 0.1947249 
#> target_dist - cur_dist: 0.1847249 
#> target_dist - cur_dist: 0.1747249 
#> target_dist - cur_dist: 0.1647249 
#> target_dist - cur_dist: 0.1547249 
#> target_dist - cur_dist: 0.1447249 
#> target_dist - cur_dist: 0.1347249 
#> target_dist - cur_dist: 0.1247249 
#> target_dist - cur_dist: 0.1147249 
#> target_dist - cur_dist: 0.1047249 
#> target_dist - cur_dist: 0.09472494 
#> target_dist - cur_dist: 0.08472494 
#> target_dist - cur_dist: 0.07472494 
#> target_dist - cur_dist: 0.06472494 
#> target_dist - cur_dist: 0.05472494 
#> target_dist - cur_dist: 0.04472494 
#> target_dist - cur_dist: 0.03472494 
#> target_dist - cur_dist: 0.02472494 
#> target_dist - cur_dist: 0.01472494 
#> target_dist - cur_dist: 0.004724942 
#> generation:  dist =   1.370737 
#> target_dist - cur_dist: 1.370737 
#> target_dist - cur_dist: 1.360737 
#> target_dist - cur_dist: 1.350737 
#> target_dist - cur_dist: 1.340737 
#> target_dist - cur_dist: 1.330737 
#> target_dist - cur_dist: 1.320737 
#> target_dist - cur_dist: 1.310737 
#> target_dist - cur_dist: 1.300737 
#> target_dist - cur_dist: 1.290737 
#> target_dist - cur_dist: 1.280737 
#> target_dist - cur_dist: 1.270737 
#> target_dist - cur_dist: 1.260737 
#> target_dist - cur_dist: 1.250737 
#> target_dist - cur_dist: 1.240737 
#> target_dist - cur_dist: 1.230737 
#> target_dist - cur_dist: 1.220737 
#> target_dist - cur_dist: 1.210737 
#> target_dist - cur_dist: 1.200737 
#> target_dist - cur_dist: 1.190737 
#> target_dist - cur_dist: 1.180737 
#> target_dist - cur_dist: 1.170737 
#> target_dist - cur_dist: 1.160737 
#> target_dist - cur_dist: 1.150737 
#> target_dist - cur_dist: 1.140737 
#> target_dist - cur_dist: 1.130737 
#> target_dist - cur_dist: 1.120737 
#> target_dist - cur_dist: 1.110737 
#> target_dist - cur_dist: 1.100737 
#> target_dist - cur_dist: 1.090737 
#> target_dist - cur_dist: 1.080737 
#> target_dist - cur_dist: 1.070737 
#> target_dist - cur_dist: 1.060737 
#> target_dist - cur_dist: 1.050737 
#> target_dist - cur_dist: 1.040737 
#> target_dist - cur_dist: 1.030737 
#> target_dist - cur_dist: 1.020737 
#> target_dist - cur_dist: 1.010737 
#> target_dist - cur_dist: 1.000737 
#> target_dist - cur_dist: 0.9907369 
#> target_dist - cur_dist: 0.9807369 
#> target_dist - cur_dist: 0.9707369 
#> target_dist - cur_dist: 0.9607369 
#> target_dist - cur_dist: 0.9507369 
#> target_dist - cur_dist: 0.9407369 
#> target_dist - cur_dist: 0.9307369 
#> target_dist - cur_dist: 0.9207369 
#> target_dist - cur_dist: 0.9107369 
#> target_dist - cur_dist: 0.9007369 
#> target_dist - cur_dist: 0.8907369 
#> target_dist - cur_dist: 0.8807369 
#> target_dist - cur_dist: 0.8707369 
#> target_dist - cur_dist: 0.8607369 
#> target_dist - cur_dist: 0.8507369 
#> target_dist - cur_dist: 0.8407369 
#> target_dist - cur_dist: 0.8307369 
#> target_dist - cur_dist: 0.8207369 
#> target_dist - cur_dist: 0.8107369 
#> target_dist - cur_dist: 0.8007369 
#> target_dist - cur_dist: 0.7907369 
#> target_dist - cur_dist: 0.7807369 
#> target_dist - cur_dist: 0.7707369 
#> target_dist - cur_dist: 0.7607369 
#> target_dist - cur_dist: 0.7507369 
#> target_dist - cur_dist: 0.7407369 
#> target_dist - cur_dist: 0.7307369 
#> target_dist - cur_dist: 0.7207369 
#> target_dist - cur_dist: 0.7107369 
#> target_dist - cur_dist: 0.7007369 
#> target_dist - cur_dist: 0.6907369 
#> target_dist - cur_dist: 0.6807369 
#> target_dist - cur_dist: 0.6707369 
#> target_dist - cur_dist: 0.6607369 
#> target_dist - cur_dist: 0.6507369 
#> target_dist - cur_dist: 0.6407369 
#> target_dist - cur_dist: 0.6307369 
#> target_dist - cur_dist: 0.6207369 
#> target_dist - cur_dist: 0.6107369 
#> target_dist - cur_dist: 0.6007369 
#> target_dist - cur_dist: 0.5907369 
#> target_dist - cur_dist: 0.5807369 
#> target_dist - cur_dist: 0.5707369 
#> target_dist - cur_dist: 0.5607369 
#> target_dist - cur_dist: 0.5507369 
#> target_dist - cur_dist: 0.5407369 
#> target_dist - cur_dist: 0.5307369 
#> target_dist - cur_dist: 0.5207369 
#> target_dist - cur_dist: 0.5107369 
#> target_dist - cur_dist: 0.5007369 
#> target_dist - cur_dist: 0.4907369 
#> target_dist - cur_dist: 0.4807369 
#> target_dist - cur_dist: 0.4707369 
#> target_dist - cur_dist: 0.4607369 
#> target_dist - cur_dist: 0.4507369 
#> target_dist - cur_dist: 0.4407369 
#> target_dist - cur_dist: 0.4307369 
#> target_dist - cur_dist: 0.4207369 
#> target_dist - cur_dist: 0.4107369 
#> target_dist - cur_dist: 0.4007369 
#> target_dist - cur_dist: 0.3907369 
#> target_dist - cur_dist: 0.3807369 
#> target_dist - cur_dist: 0.3707369 
#> target_dist - cur_dist: 0.3607369 
#> target_dist - cur_dist: 0.3507369 
#> target_dist - cur_dist: 0.3407369 
#> target_dist - cur_dist: 0.3307369 
#> target_dist - cur_dist: 0.3207369 
#> target_dist - cur_dist: 0.3107369 
#> target_dist - cur_dist: 0.3007369 
#> target_dist - cur_dist: 0.2907369 
#> target_dist - cur_dist: 0.2807369 
#> target_dist - cur_dist: 0.2707369 
#> target_dist - cur_dist: 0.2607369 
#> target_dist - cur_dist: 0.2507369 
#> target_dist - cur_dist: 0.2407369 
#> target_dist - cur_dist: 0.2307369 
#> target_dist - cur_dist: 0.2207369 
#> target_dist - cur_dist: 0.2107369 
#> target_dist - cur_dist: 0.2007369 
#> target_dist - cur_dist: 0.1907369 
#> target_dist - cur_dist: 0.1807369 
#> target_dist - cur_dist: 0.1707369 
#> target_dist - cur_dist: 0.1607369 
#> target_dist - cur_dist: 0.1507369 
#> target_dist - cur_dist: 0.1407369 
#> target_dist - cur_dist: 0.1307369 
#> target_dist - cur_dist: 0.1207369 
#> target_dist - cur_dist: 0.1107369 
#> target_dist - cur_dist: 0.1007369 
#> target_dist - cur_dist: 0.09073689 
#> target_dist - cur_dist: 0.08073689 
#> target_dist - cur_dist: 0.07073689 
#> target_dist - cur_dist: 0.06073689 
#> target_dist - cur_dist: 0.05073689 
#> target_dist - cur_dist: 0.04073689 
#> target_dist - cur_dist: 0.03073689 
#> target_dist - cur_dist: 0.02073689 
#> target_dist - cur_dist: 0.01073689 
#> target_dist - cur_dist: 0.0007368913 
#> generation:  dist =   1.219314 
#> target_dist - cur_dist: 1.219314 
#> target_dist - cur_dist: 1.209314 
#> target_dist - cur_dist: 1.199314 
#> target_dist - cur_dist: 1.189314 
#> target_dist - cur_dist: 1.179314 
#> target_dist - cur_dist: 1.169314 
#> target_dist - cur_dist: 1.159314 
#> target_dist - cur_dist: 1.149314 
#> target_dist - cur_dist: 1.139314 
#> target_dist - cur_dist: 1.129314 
#> target_dist - cur_dist: 1.119314 
#> target_dist - cur_dist: 1.109314 
#> target_dist - cur_dist: 1.099314 
#> target_dist - cur_dist: 1.089314 
#> target_dist - cur_dist: 1.079314 
#> target_dist - cur_dist: 1.069314 
#> target_dist - cur_dist: 1.059314 
#> target_dist - cur_dist: 1.049314 
#> target_dist - cur_dist: 1.039314 
#> target_dist - cur_dist: 1.029314 
#> target_dist - cur_dist: 1.019314 
#> target_dist - cur_dist: 1.009314 
#> target_dist - cur_dist: 0.9993137 
#> target_dist - cur_dist: 0.9893137 
#> target_dist - cur_dist: 0.9793137 
#> target_dist - cur_dist: 0.9693137 
#> target_dist - cur_dist: 0.9593137 
#> target_dist - cur_dist: 0.9493137 
#> target_dist - cur_dist: 0.9393137 
#> target_dist - cur_dist: 0.9293137 
#> target_dist - cur_dist: 0.9193137 
#> target_dist - cur_dist: 0.9093137 
#> target_dist - cur_dist: 0.8993137 
#> target_dist - cur_dist: 0.8893137 
#> target_dist - cur_dist: 0.8793137 
#> target_dist - cur_dist: 0.8693137 
#> target_dist - cur_dist: 0.8593137 
#> target_dist - cur_dist: 0.8493137 
#> target_dist - cur_dist: 0.8393137 
#> target_dist - cur_dist: 0.8293137 
#> target_dist - cur_dist: 0.8193137 
#> target_dist - cur_dist: 0.8093137 
#> target_dist - cur_dist: 0.7993137 
#> target_dist - cur_dist: 0.7893137 
#> target_dist - cur_dist: 0.7793137 
#> target_dist - cur_dist: 0.7693137 
#> target_dist - cur_dist: 0.7593137 
#> target_dist - cur_dist: 0.7493137 
#> target_dist - cur_dist: 0.7393137 
#> target_dist - cur_dist: 0.7293137 
#> target_dist - cur_dist: 0.7193137 
#> target_dist - cur_dist: 0.7093137 
#> target_dist - cur_dist: 0.6993137 
#> target_dist - cur_dist: 0.6893137 
#> target_dist - cur_dist: 0.6793137 
#> target_dist - cur_dist: 0.6693137 
#> target_dist - cur_dist: 0.6593137 
#> target_dist - cur_dist: 0.6493137 
#> target_dist - cur_dist: 0.6393137 
#> target_dist - cur_dist: 0.6293137 
#> target_dist - cur_dist: 0.6193137 
#> target_dist - cur_dist: 0.6093137 
#> target_dist - cur_dist: 0.5993137 
#> target_dist - cur_dist: 0.5893137 
#> target_dist - cur_dist: 0.5793137 
#> target_dist - cur_dist: 0.5693137 
#> target_dist - cur_dist: 0.5593137 
#> target_dist - cur_dist: 0.5493137 
#> target_dist - cur_dist: 0.5393137 
#> target_dist - cur_dist: 0.5293137 
#> target_dist - cur_dist: 0.5193137 
#> target_dist - cur_dist: 0.5093137 
#> target_dist - cur_dist: 0.4993137 
#> target_dist - cur_dist: 0.4893137 
#> target_dist - cur_dist: 0.4793137 
#> target_dist - cur_dist: 0.4693137 
#> target_dist - cur_dist: 0.4593137 
#> target_dist - cur_dist: 0.4493137 
#> target_dist - cur_dist: 0.4393137 
#> target_dist - cur_dist: 0.4293137 
#> target_dist - cur_dist: 0.4193137 
#> target_dist - cur_dist: 0.4093137 
#> target_dist - cur_dist: 0.3993137 
#> target_dist - cur_dist: 0.3893137 
#> target_dist - cur_dist: 0.3793137 
#> target_dist - cur_dist: 0.3693137 
#> target_dist - cur_dist: 0.3593137 
#> target_dist - cur_dist: 0.3493137 
#> target_dist - cur_dist: 0.3393137 
#> target_dist - cur_dist: 0.3293137 
#> target_dist - cur_dist: 0.3193137 
#> target_dist - cur_dist: 0.3093137 
#> target_dist - cur_dist: 0.2993137 
#> target_dist - cur_dist: 0.2893137 
#> target_dist - cur_dist: 0.2793137 
#> target_dist - cur_dist: 0.2693137 
#> target_dist - cur_dist: 0.2593137 
#> target_dist - cur_dist: 0.2493137 
#> target_dist - cur_dist: 0.2393137 
#> target_dist - cur_dist: 0.2293137 
#> target_dist - cur_dist: 0.2193137 
#> target_dist - cur_dist: 0.2093137 
#> target_dist - cur_dist: 0.1993137 
#> target_dist - cur_dist: 0.1893137 
#> target_dist - cur_dist: 0.1793137 
#> target_dist - cur_dist: 0.1693137 
#> target_dist - cur_dist: 0.1593137 
#> target_dist - cur_dist: 0.1493137 
#> target_dist - cur_dist: 0.1393137 
#> target_dist - cur_dist: 0.1293137 
#> target_dist - cur_dist: 0.1193137 
#> target_dist - cur_dist: 0.1093137 
#> target_dist - cur_dist: 0.0993137 
#> target_dist - cur_dist: 0.0893137 
#> target_dist - cur_dist: 0.0793137 
#> target_dist - cur_dist: 0.0693137 
#> target_dist - cur_dist: 0.0593137 
#> target_dist - cur_dist: 0.0493137 
#> target_dist - cur_dist: 0.0393137 
#> target_dist - cur_dist: 0.0293137 
#> target_dist - cur_dist: 0.0193137 
#> target_dist - cur_dist: 0.009313698 
#> generation:  dist =   1.290812 
#> target_dist - cur_dist: 1.290812 
#> target_dist - cur_dist: 1.280812 
#> target_dist - cur_dist: 1.270812 
#> target_dist - cur_dist: 1.260812 
#> target_dist - cur_dist: 1.250812 
#> target_dist - cur_dist: 1.240812 
#> target_dist - cur_dist: 1.230812 
#> target_dist - cur_dist: 1.220812 
#> target_dist - cur_dist: 1.210812 
#> target_dist - cur_dist: 1.200812 
#> target_dist - cur_dist: 1.190812 
#> target_dist - cur_dist: 1.180812 
#> target_dist - cur_dist: 1.170812 
#> target_dist - cur_dist: 1.160812 
#> target_dist - cur_dist: 1.150812 
#> target_dist - cur_dist: 1.140812 
#> target_dist - cur_dist: 1.130812 
#> target_dist - cur_dist: 1.120812 
#> target_dist - cur_dist: 1.110812 
#> target_dist - cur_dist: 1.100812 
#> target_dist - cur_dist: 1.090812 
#> target_dist - cur_dist: 1.080812 
#> target_dist - cur_dist: 1.070812 
#> target_dist - cur_dist: 1.060812 
#> target_dist - cur_dist: 1.050812 
#> target_dist - cur_dist: 1.040812 
#> target_dist - cur_dist: 1.030812 
#> target_dist - cur_dist: 1.020812 
#> target_dist - cur_dist: 1.010812 
#> target_dist - cur_dist: 1.000812 
#> target_dist - cur_dist: 0.9908119 
#> target_dist - cur_dist: 0.9808119 
#> target_dist - cur_dist: 0.9708119 
#> target_dist - cur_dist: 0.9608119 
#> target_dist - cur_dist: 0.9508119 
#> target_dist - cur_dist: 0.9408119 
#> target_dist - cur_dist: 0.9308119 
#> target_dist - cur_dist: 0.9208119 
#> target_dist - cur_dist: 0.9108119 
#> target_dist - cur_dist: 0.9008119 
#> target_dist - cur_dist: 0.8908119 
#> target_dist - cur_dist: 0.8808119 
#> target_dist - cur_dist: 0.8708119 
#> target_dist - cur_dist: 0.8608119 
#> target_dist - cur_dist: 0.8508119 
#> target_dist - cur_dist: 0.8408119 
#> target_dist - cur_dist: 0.8308119 
#> target_dist - cur_dist: 0.8208119 
#> target_dist - cur_dist: 0.8108119 
#> target_dist - cur_dist: 0.8008119 
#> target_dist - cur_dist: 0.7908119 
#> target_dist - cur_dist: 0.7808119 
#> target_dist - cur_dist: 0.7708119 
#> target_dist - cur_dist: 0.7608119 
#> target_dist - cur_dist: 0.7508119 
#> target_dist - cur_dist: 0.7408119 
#> target_dist - cur_dist: 0.7308119 
#> target_dist - cur_dist: 0.7208119 
#> target_dist - cur_dist: 0.7108119 
#> target_dist - cur_dist: 0.7008119 
#> target_dist - cur_dist: 0.6908119 
#> target_dist - cur_dist: 0.6808119 
#> target_dist - cur_dist: 0.6708119 
#> target_dist - cur_dist: 0.6608119 
#> target_dist - cur_dist: 0.6508119 
#> target_dist - cur_dist: 0.6408119 
#> target_dist - cur_dist: 0.6308119 
#> target_dist - cur_dist: 0.6208119 
#> target_dist - cur_dist: 0.6108119 
#> target_dist - cur_dist: 0.6008119 
#> target_dist - cur_dist: 0.5908119 
#> target_dist - cur_dist: 0.5808119 
#> target_dist - cur_dist: 0.5708119 
#> target_dist - cur_dist: 0.5608119 
#> target_dist - cur_dist: 0.5508119 
#> target_dist - cur_dist: 0.5408119 
#> target_dist - cur_dist: 0.5308119 
#> target_dist - cur_dist: 0.5208119 
#> target_dist - cur_dist: 0.5108119 
#> target_dist - cur_dist: 0.5008119 
#> target_dist - cur_dist: 0.4908119 
#> target_dist - cur_dist: 0.4808119 
#> target_dist - cur_dist: 0.4708119 
#> target_dist - cur_dist: 0.4608119 
#> target_dist - cur_dist: 0.4508119 
#> target_dist - cur_dist: 0.4408119 
#> target_dist - cur_dist: 0.4308119 
#> target_dist - cur_dist: 0.4208119 
#> target_dist - cur_dist: 0.4108119 
#> target_dist - cur_dist: 0.4008119 
#> target_dist - cur_dist: 0.3908119 
#> target_dist - cur_dist: 0.3808119 
#> target_dist - cur_dist: 0.3708119 
#> target_dist - cur_dist: 0.3608119 
#> target_dist - cur_dist: 0.3508119 
#> target_dist - cur_dist: 0.3408119 
#> target_dist - cur_dist: 0.3308119 
#> target_dist - cur_dist: 0.3208119 
#> target_dist - cur_dist: 0.3108119 
#> target_dist - cur_dist: 0.3008119 
#> target_dist - cur_dist: 0.2908119 
#> target_dist - cur_dist: 0.2808119 
#> target_dist - cur_dist: 0.2708119 
#> target_dist - cur_dist: 0.2608119 
#> target_dist - cur_dist: 0.2508119 
#> target_dist - cur_dist: 0.2408119 
#> target_dist - cur_dist: 0.2308119 
#> target_dist - cur_dist: 0.2208119 
#> target_dist - cur_dist: 0.2108119 
#> target_dist - cur_dist: 0.2008119 
#> target_dist - cur_dist: 0.1908119 
#> target_dist - cur_dist: 0.1808119 
#> target_dist - cur_dist: 0.1708119 
#> target_dist - cur_dist: 0.1608119 
#> target_dist - cur_dist: 0.1508119 
#> target_dist - cur_dist: 0.1408119 
#> target_dist - cur_dist: 0.1308119 
#> target_dist - cur_dist: 0.1208119 
#> target_dist - cur_dist: 0.1108119 
#> target_dist - cur_dist: 0.1008119 
#> target_dist - cur_dist: 0.09081192 
#> target_dist - cur_dist: 0.08081192 
#> target_dist - cur_dist: 0.07081192 
#> target_dist - cur_dist: 0.06081192 
#> target_dist - cur_dist: 0.05081192 
#> target_dist - cur_dist: 0.04081192 
#> target_dist - cur_dist: 0.03081192 
#> target_dist - cur_dist: 0.02081192 
#> target_dist - cur_dist: 0.01081192 
#> target_dist - cur_dist: 0.0008119152 
#> generation:  dist =   1.394297 
#> target_dist - cur_dist: 1.394297 
#> target_dist - cur_dist: 1.384297 
#> target_dist - cur_dist: 1.374297 
#> target_dist - cur_dist: 1.364297 
#> target_dist - cur_dist: 1.354297 
#> target_dist - cur_dist: 1.344297 
#> target_dist - cur_dist: 1.334297 
#> target_dist - cur_dist: 1.324297 
#> target_dist - cur_dist: 1.314297 
#> target_dist - cur_dist: 1.304297 
#> target_dist - cur_dist: 1.294297 
#> target_dist - cur_dist: 1.284297 
#> target_dist - cur_dist: 1.274297 
#> target_dist - cur_dist: 1.264297 
#> target_dist - cur_dist: 1.254297 
#> target_dist - cur_dist: 1.244297 
#> target_dist - cur_dist: 1.234297 
#> target_dist - cur_dist: 1.224297 
#> target_dist - cur_dist: 1.214297 
#> target_dist - cur_dist: 1.204297 
#> target_dist - cur_dist: 1.194297 
#> target_dist - cur_dist: 1.184297 
#> target_dist - cur_dist: 1.174297 
#> target_dist - cur_dist: 1.164297 
#> target_dist - cur_dist: 1.154297 
#> target_dist - cur_dist: 1.144297 
#> target_dist - cur_dist: 1.134297 
#> target_dist - cur_dist: 1.124297 
#> target_dist - cur_dist: 1.114297 
#> target_dist - cur_dist: 1.104297 
#> target_dist - cur_dist: 1.094297 
#> target_dist - cur_dist: 1.084297 
#> target_dist - cur_dist: 1.074297 
#> target_dist - cur_dist: 1.064297 
#> target_dist - cur_dist: 1.054297 
#> target_dist - cur_dist: 1.044297 
#> target_dist - cur_dist: 1.034297 
#> target_dist - cur_dist: 1.024297 
#> target_dist - cur_dist: 1.014297 
#> target_dist - cur_dist: 1.004297 
#> target_dist - cur_dist: 0.9942971 
#> target_dist - cur_dist: 0.9842971 
#> target_dist - cur_dist: 0.9742971 
#> target_dist - cur_dist: 0.9642971 
#> target_dist - cur_dist: 0.9542971 
#> target_dist - cur_dist: 0.9442971 
#> target_dist - cur_dist: 0.9342971 
#> target_dist - cur_dist: 0.9242971 
#> target_dist - cur_dist: 0.9142971 
#> target_dist - cur_dist: 0.9042971 
#> target_dist - cur_dist: 0.8942971 
#> target_dist - cur_dist: 0.8842971 
#> target_dist - cur_dist: 0.8742971 
#> target_dist - cur_dist: 0.8642971 
#> target_dist - cur_dist: 0.8542971 
#> target_dist - cur_dist: 0.8442971 
#> target_dist - cur_dist: 0.8342971 
#> target_dist - cur_dist: 0.8242971 
#> target_dist - cur_dist: 0.8142971 
#> target_dist - cur_dist: 0.8042971 
#> target_dist - cur_dist: 0.7942971 
#> target_dist - cur_dist: 0.7842971 
#> target_dist - cur_dist: 0.7742971 
#> target_dist - cur_dist: 0.7642971 
#> target_dist - cur_dist: 0.7542971 
#> target_dist - cur_dist: 0.7442971 
#> target_dist - cur_dist: 0.7342971 
#> target_dist - cur_dist: 0.7242971 
#> target_dist - cur_dist: 0.7142971 
#> target_dist - cur_dist: 0.7042971 
#> target_dist - cur_dist: 0.6942971 
#> target_dist - cur_dist: 0.6842971 
#> target_dist - cur_dist: 0.6742971 
#> target_dist - cur_dist: 0.6642971 
#> target_dist - cur_dist: 0.6542971 
#> target_dist - cur_dist: 0.6442971 
#> target_dist - cur_dist: 0.6342971 
#> target_dist - cur_dist: 0.6242971 
#> target_dist - cur_dist: 0.6142971 
#> target_dist - cur_dist: 0.6042971 
#> target_dist - cur_dist: 0.5942971 
#> target_dist - cur_dist: 0.5842971 
#> target_dist - cur_dist: 0.5742971 
#> target_dist - cur_dist: 0.5642971 
#> target_dist - cur_dist: 0.5542971 
#> target_dist - cur_dist: 0.5442971 
#> target_dist - cur_dist: 0.5342971 
#> target_dist - cur_dist: 0.5242971 
#> target_dist - cur_dist: 0.5142971 
#> target_dist - cur_dist: 0.5042971 
#> target_dist - cur_dist: 0.4942971 
#> target_dist - cur_dist: 0.4842971 
#> target_dist - cur_dist: 0.4742971 
#> target_dist - cur_dist: 0.4642971 
#> target_dist - cur_dist: 0.4542971 
#> target_dist - cur_dist: 0.4442971 
#> target_dist - cur_dist: 0.4342971 
#> target_dist - cur_dist: 0.4242971 
#> target_dist - cur_dist: 0.4142971 
#> target_dist - cur_dist: 0.4042971 
#> target_dist - cur_dist: 0.3942971 
#> target_dist - cur_dist: 0.3842971 
#> target_dist - cur_dist: 0.3742971 
#> target_dist - cur_dist: 0.3642971 
#> target_dist - cur_dist: 0.3542971 
#> target_dist - cur_dist: 0.3442971 
#> target_dist - cur_dist: 0.3342971 
#> target_dist - cur_dist: 0.3242971 
#> target_dist - cur_dist: 0.3142971 
#> target_dist - cur_dist: 0.3042971 
#> target_dist - cur_dist: 0.2942971 
#> target_dist - cur_dist: 0.2842971 
#> target_dist - cur_dist: 0.2742971 
#> target_dist - cur_dist: 0.2642971 
#> target_dist - cur_dist: 0.2542971 
#> target_dist - cur_dist: 0.2442971 
#> target_dist - cur_dist: 0.2342971 
#> target_dist - cur_dist: 0.2242971 
#> target_dist - cur_dist: 0.2142971 
#> target_dist - cur_dist: 0.2042971 
#> target_dist - cur_dist: 0.1942971 
#> target_dist - cur_dist: 0.1842971 
#> target_dist - cur_dist: 0.1742971 
#> target_dist - cur_dist: 0.1642971 
#> target_dist - cur_dist: 0.1542971 
#> target_dist - cur_dist: 0.1442971 
#> target_dist - cur_dist: 0.1342971 
#> target_dist - cur_dist: 0.1242971 
#> target_dist - cur_dist: 0.1142971 
#> target_dist - cur_dist: 0.1042971 
#> target_dist - cur_dist: 0.09429712 
#> target_dist - cur_dist: 0.08429712 
#> target_dist - cur_dist: 0.07429712 
#> target_dist - cur_dist: 0.06429712 
#> target_dist - cur_dist: 0.05429712 
#> target_dist - cur_dist: 0.04429712 
#> target_dist - cur_dist: 0.03429712 
#> target_dist - cur_dist: 0.02429712 
#> target_dist - cur_dist: 0.01429712 
#> target_dist - cur_dist: 0.004297118 
#> generation:  dist =   1.011288 
#> target_dist - cur_dist: 1.011288 
#> target_dist - cur_dist: 1.001288 
#> target_dist - cur_dist: 0.9912877 
#> target_dist - cur_dist: 0.9812877 
#> target_dist - cur_dist: 0.9712877 
#> target_dist - cur_dist: 0.9612877 
#> target_dist - cur_dist: 0.9512877 
#> target_dist - cur_dist: 0.9412877 
#> target_dist - cur_dist: 0.9312877 
#> target_dist - cur_dist: 0.9212877 
#> target_dist - cur_dist: 0.9112877 
#> target_dist - cur_dist: 0.9012877 
#> target_dist - cur_dist: 0.8912877 
#> target_dist - cur_dist: 0.8812877 
#> target_dist - cur_dist: 0.8712877 
#> target_dist - cur_dist: 0.8612877 
#> target_dist - cur_dist: 0.8512877 
#> target_dist - cur_dist: 0.8412877 
#> target_dist - cur_dist: 0.8312877 
#> target_dist - cur_dist: 0.8212877 
#> target_dist - cur_dist: 0.8112877 
#> target_dist - cur_dist: 0.8012877 
#> target_dist - cur_dist: 0.7912877 
#> target_dist - cur_dist: 0.7812877 
#> target_dist - cur_dist: 0.7712877 
#> target_dist - cur_dist: 0.7612877 
#> target_dist - cur_dist: 0.7512877 
#> target_dist - cur_dist: 0.7412877 
#> target_dist - cur_dist: 0.7312877 
#> target_dist - cur_dist: 0.7212877 
#> target_dist - cur_dist: 0.7112877 
#> target_dist - cur_dist: 0.7012877 
#> target_dist - cur_dist: 0.6912877 
#> target_dist - cur_dist: 0.6812877 
#> target_dist - cur_dist: 0.6712877 
#> target_dist - cur_dist: 0.6612877 
#> target_dist - cur_dist: 0.6512877 
#> target_dist - cur_dist: 0.6412877 
#> target_dist - cur_dist: 0.6312877 
#> target_dist - cur_dist: 0.6212877 
#> target_dist - cur_dist: 0.6112877 
#> target_dist - cur_dist: 0.6012877 
#> target_dist - cur_dist: 0.5912877 
#> target_dist - cur_dist: 0.5812877 
#> target_dist - cur_dist: 0.5712877 
#> target_dist - cur_dist: 0.5612877 
#> target_dist - cur_dist: 0.5512877 
#> target_dist - cur_dist: 0.5412877 
#> target_dist - cur_dist: 0.5312877 
#> target_dist - cur_dist: 0.5212877 
#> target_dist - cur_dist: 0.5112877 
#> target_dist - cur_dist: 0.5012877 
#> target_dist - cur_dist: 0.4912877 
#> target_dist - cur_dist: 0.4812877 
#> target_dist - cur_dist: 0.4712877 
#> target_dist - cur_dist: 0.4612877 
#> target_dist - cur_dist: 0.4512877 
#> target_dist - cur_dist: 0.4412877 
#> target_dist - cur_dist: 0.4312877 
#> target_dist - cur_dist: 0.4212877 
#> target_dist - cur_dist: 0.4112877 
#> target_dist - cur_dist: 0.4012877 
#> target_dist - cur_dist: 0.3912877 
#> target_dist - cur_dist: 0.3812877 
#> target_dist - cur_dist: 0.3712877 
#> target_dist - cur_dist: 0.3612877 
#> target_dist - cur_dist: 0.3512877 
#> target_dist - cur_dist: 0.3412877 
#> target_dist - cur_dist: 0.3312877 
#> target_dist - cur_dist: 0.3212877 
#> target_dist - cur_dist: 0.3112877 
#> target_dist - cur_dist: 0.3012877 
#> target_dist - cur_dist: 0.2912877 
#> target_dist - cur_dist: 0.2812877 
#> target_dist - cur_dist: 0.2712877 
#> target_dist - cur_dist: 0.2612877 
#> target_dist - cur_dist: 0.2512877 
#> target_dist - cur_dist: 0.2412877 
#> target_dist - cur_dist: 0.2312877 
#> target_dist - cur_dist: 0.2212877 
#> target_dist - cur_dist: 0.2112877 
#> target_dist - cur_dist: 0.2012877 
#> target_dist - cur_dist: 0.1912877 
#> target_dist - cur_dist: 0.1812877 
#> target_dist - cur_dist: 0.1712877 
#> target_dist - cur_dist: 0.1612877 
#> target_dist - cur_dist: 0.1512877 
#> target_dist - cur_dist: 0.1412877 
#> target_dist - cur_dist: 0.1312877 
#> target_dist - cur_dist: 0.1212877 
#> target_dist - cur_dist: 0.1112877 
#> target_dist - cur_dist: 0.1012877 
#> target_dist - cur_dist: 0.09128767 
#> target_dist - cur_dist: 0.08128767 
#> target_dist - cur_dist: 0.07128767 
#> target_dist - cur_dist: 0.06128767 
#> target_dist - cur_dist: 0.05128767 
#> target_dist - cur_dist: 0.04128767 
#> target_dist - cur_dist: 0.03128767 
#> target_dist - cur_dist: 0.02128767 
#> target_dist - cur_dist: 0.01128767 
#> target_dist - cur_dist: 0.001287667 
#> generation:  dist =   1.291949 
#> target_dist - cur_dist: 1.291949 
#> target_dist - cur_dist: 1.281949 
#> target_dist - cur_dist: 1.271949 
#> target_dist - cur_dist: 1.261949 
#> target_dist - cur_dist: 1.251949 
#> target_dist - cur_dist: 1.241949 
#> target_dist - cur_dist: 1.231949 
#> target_dist - cur_dist: 1.221949 
#> target_dist - cur_dist: 1.211949 
#> target_dist - cur_dist: 1.201949 
#> target_dist - cur_dist: 1.191949 
#> target_dist - cur_dist: 1.181949 
#> target_dist - cur_dist: 1.171949 
#> target_dist - cur_dist: 1.161949 
#> target_dist - cur_dist: 1.151949 
#> target_dist - cur_dist: 1.141949 
#> target_dist - cur_dist: 1.131949 
#> target_dist - cur_dist: 1.121949 
#> target_dist - cur_dist: 1.111949 
#> target_dist - cur_dist: 1.101949 
#> target_dist - cur_dist: 1.091949 
#> target_dist - cur_dist: 1.081949 
#> target_dist - cur_dist: 1.071949 
#> target_dist - cur_dist: 1.061949 
#> target_dist - cur_dist: 1.051949 
#> target_dist - cur_dist: 1.041949 
#> target_dist - cur_dist: 1.031949 
#> target_dist - cur_dist: 1.021949 
#> target_dist - cur_dist: 1.011949 
#> target_dist - cur_dist: 1.001949 
#> target_dist - cur_dist: 0.9919487 
#> target_dist - cur_dist: 0.9819487 
#> target_dist - cur_dist: 0.9719487 
#> target_dist - cur_dist: 0.9619487 
#> target_dist - cur_dist: 0.9519487 
#> target_dist - cur_dist: 0.9419487 
#> target_dist - cur_dist: 0.9319487 
#> target_dist - cur_dist: 0.9219487 
#> target_dist - cur_dist: 0.9119487 
#> target_dist - cur_dist: 0.9019487 
#> target_dist - cur_dist: 0.8919487 
#> target_dist - cur_dist: 0.8819487 
#> target_dist - cur_dist: 0.8719487 
#> target_dist - cur_dist: 0.8619487 
#> target_dist - cur_dist: 0.8519487 
#> target_dist - cur_dist: 0.8419487 
#> target_dist - cur_dist: 0.8319487 
#> target_dist - cur_dist: 0.8219487 
#> target_dist - cur_dist: 0.8119487 
#> target_dist - cur_dist: 0.8019487 
#> target_dist - cur_dist: 0.7919487 
#> target_dist - cur_dist: 0.7819487 
#> target_dist - cur_dist: 0.7719487 
#> target_dist - cur_dist: 0.7619487 
#> target_dist - cur_dist: 0.7519487 
#> target_dist - cur_dist: 0.7419487 
#> target_dist - cur_dist: 0.7319487 
#> target_dist - cur_dist: 0.7219487 
#> target_dist - cur_dist: 0.7119487 
#> target_dist - cur_dist: 0.7019487 
#> target_dist - cur_dist: 0.6919487 
#> target_dist - cur_dist: 0.6819487 
#> target_dist - cur_dist: 0.6719487 
#> target_dist - cur_dist: 0.6619487 
#> target_dist - cur_dist: 0.6519487 
#> target_dist - cur_dist: 0.6419487 
#> target_dist - cur_dist: 0.6319487 
#> target_dist - cur_dist: 0.6219487 
#> target_dist - cur_dist: 0.6119487 
#> target_dist - cur_dist: 0.6019487 
#> target_dist - cur_dist: 0.5919487 
#> target_dist - cur_dist: 0.5819487 
#> target_dist - cur_dist: 0.5719487 
#> target_dist - cur_dist: 0.5619487 
#> target_dist - cur_dist: 0.5519487 
#> target_dist - cur_dist: 0.5419487 
#> target_dist - cur_dist: 0.5319487 
#> target_dist - cur_dist: 0.5219487 
#> target_dist - cur_dist: 0.5119487 
#> target_dist - cur_dist: 0.5019487 
#> target_dist - cur_dist: 0.4919487 
#> target_dist - cur_dist: 0.4819487 
#> target_dist - cur_dist: 0.4719487 
#> target_dist - cur_dist: 0.4619487 
#> target_dist - cur_dist: 0.4519487 
#> target_dist - cur_dist: 0.4419487 
#> target_dist - cur_dist: 0.4319487 
#> target_dist - cur_dist: 0.4219487 
#> target_dist - cur_dist: 0.4119487 
#> target_dist - cur_dist: 0.4019487 
#> target_dist - cur_dist: 0.3919487 
#> target_dist - cur_dist: 0.3819487 
#> target_dist - cur_dist: 0.3719487 
#> target_dist - cur_dist: 0.3619487 
#> target_dist - cur_dist: 0.3519487 
#> target_dist - cur_dist: 0.3419487 
#> target_dist - cur_dist: 0.3319487 
#> target_dist - cur_dist: 0.3219487 
#> target_dist - cur_dist: 0.3119487 
#> target_dist - cur_dist: 0.3019487 
#> target_dist - cur_dist: 0.2919487 
#> target_dist - cur_dist: 0.2819487 
#> target_dist - cur_dist: 0.2719487 
#> target_dist - cur_dist: 0.2619487 
#> target_dist - cur_dist: 0.2519487 
#> target_dist - cur_dist: 0.2419487 
#> target_dist - cur_dist: 0.2319487 
#> target_dist - cur_dist: 0.2219487 
#> target_dist - cur_dist: 0.2119487 
#> target_dist - cur_dist: 0.2019487 
#> target_dist - cur_dist: 0.1919487 
#> target_dist - cur_dist: 0.1819487 
#> target_dist - cur_dist: 0.1719487 
#> target_dist - cur_dist: 0.1619487 
#> target_dist - cur_dist: 0.1519487 
#> target_dist - cur_dist: 0.1419487 
#> target_dist - cur_dist: 0.1319487 
#> target_dist - cur_dist: 0.1219487 
#> target_dist - cur_dist: 0.1119487 
#> target_dist - cur_dist: 0.1019487 
#> target_dist - cur_dist: 0.09194875 
#> target_dist - cur_dist: 0.08194875 
#> target_dist - cur_dist: 0.07194875 
#> target_dist - cur_dist: 0.06194875 
#> target_dist - cur_dist: 0.05194875 
#> target_dist - cur_dist: 0.04194875 
#> target_dist - cur_dist: 0.03194875 
#> target_dist - cur_dist: 0.02194875 
#> target_dist - cur_dist: 0.01194875 
#> target_dist - cur_dist: 0.001948749 

#> [1]    6    1 1163

dim(interpolate(t1, 0.05))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.299216 
#> target_dist - cur_dist: 1.299216 
#> target_dist - cur_dist: 1.249216 
#> target_dist - cur_dist: 1.199216 
#> target_dist - cur_dist: 1.149216 
#> target_dist - cur_dist: 1.099216 
#> target_dist - cur_dist: 1.049216 
#> target_dist - cur_dist: 0.9992157 
#> target_dist - cur_dist: 0.9492157 
#> target_dist - cur_dist: 0.8992157 
#> target_dist - cur_dist: 0.8492157 
#> target_dist - cur_dist: 0.7992157 
#> target_dist - cur_dist: 0.7492157 
#> target_dist - cur_dist: 0.6992157 
#> target_dist - cur_dist: 0.6492157 
#> target_dist - cur_dist: 0.5992157 
#> target_dist - cur_dist: 0.5492157 
#> target_dist - cur_dist: 0.4992157 
#> target_dist - cur_dist: 0.4492157 
#> target_dist - cur_dist: 0.3992157 
#> target_dist - cur_dist: 0.3492157 
#> target_dist - cur_dist: 0.2992157 
#> target_dist - cur_dist: 0.2492157 
#> target_dist - cur_dist: 0.1992157 
#> target_dist - cur_dist: 0.1492157 
#> target_dist - cur_dist: 0.09921574 
#> target_dist - cur_dist: 0.04921574 
#> generation:  dist =   1.348672 
#> target_dist - cur_dist: 1.348672 
#> target_dist - cur_dist: 1.298672 
#> target_dist - cur_dist: 1.248672 
#> target_dist - cur_dist: 1.198672 
#> target_dist - cur_dist: 1.148672 
#> target_dist - cur_dist: 1.098672 
#> target_dist - cur_dist: 1.048672 
#> target_dist - cur_dist: 0.9986724 
#> target_dist - cur_dist: 0.9486724 
#> target_dist - cur_dist: 0.8986724 
#> target_dist - cur_dist: 0.8486724 
#> target_dist - cur_dist: 0.7986724 
#> target_dist - cur_dist: 0.7486724 
#> target_dist - cur_dist: 0.6986724 
#> target_dist - cur_dist: 0.6486724 
#> target_dist - cur_dist: 0.5986724 
#> target_dist - cur_dist: 0.5486724 
#> target_dist - cur_dist: 0.4986724 
#> target_dist - cur_dist: 0.4486724 
#> target_dist - cur_dist: 0.3986724 
#> target_dist - cur_dist: 0.3486724 
#> target_dist - cur_dist: 0.2986724 
#> target_dist - cur_dist: 0.2486724 
#> target_dist - cur_dist: 0.1986724 
#> target_dist - cur_dist: 0.1486724 
#> target_dist - cur_dist: 0.09867236 
#> target_dist - cur_dist: 0.04867236 
#> generation:  dist =   1.344725 
#> target_dist - cur_dist: 1.344725 
#> target_dist - cur_dist: 1.294725 
#> target_dist - cur_dist: 1.244725 
#> target_dist - cur_dist: 1.194725 
#> target_dist - cur_dist: 1.144725 
#> target_dist - cur_dist: 1.094725 
#> target_dist - cur_dist: 1.044725 
#> target_dist - cur_dist: 0.9947249 
#> target_dist - cur_dist: 0.9447249 
#> target_dist - cur_dist: 0.8947249 
#> target_dist - cur_dist: 0.8447249 
#> target_dist - cur_dist: 0.7947249 
#> target_dist - cur_dist: 0.7447249 
#> target_dist - cur_dist: 0.6947249 
#> target_dist - cur_dist: 0.6447249 
#> target_dist - cur_dist: 0.5947249 
#> target_dist - cur_dist: 0.5447249 
#> target_dist - cur_dist: 0.4947249 
#> target_dist - cur_dist: 0.4447249 
#> target_dist - cur_dist: 0.3947249 
#> target_dist - cur_dist: 0.3447249 
#> target_dist - cur_dist: 0.2947249 
#> target_dist - cur_dist: 0.2447249 
#> target_dist - cur_dist: 0.1947249 
#> target_dist - cur_dist: 0.1447249 
#> target_dist - cur_dist: 0.09472494 
#> target_dist - cur_dist: 0.04472494 
#> generation:  dist =   1.370737 
#> target_dist - cur_dist: 1.370737 
#> target_dist - cur_dist: 1.320737 
#> target_dist - cur_dist: 1.270737 
#> target_dist - cur_dist: 1.220737 
#> target_dist - cur_dist: 1.170737 
#> target_dist - cur_dist: 1.120737 
#> target_dist - cur_dist: 1.070737 
#> target_dist - cur_dist: 1.020737 
#> target_dist - cur_dist: 0.9707369 
#> target_dist - cur_dist: 0.9207369 
#> target_dist - cur_dist: 0.8707369 
#> target_dist - cur_dist: 0.8207369 
#> target_dist - cur_dist: 0.7707369 
#> target_dist - cur_dist: 0.7207369 
#> target_dist - cur_dist: 0.6707369 
#> target_dist - cur_dist: 0.6207369 
#> target_dist - cur_dist: 0.5707369 
#> target_dist - cur_dist: 0.5207369 
#> target_dist - cur_dist: 0.4707369 
#> target_dist - cur_dist: 0.4207369 
#> target_dist - cur_dist: 0.3707369 
#> target_dist - cur_dist: 0.3207369 
#> target_dist - cur_dist: 0.2707369 
#> target_dist - cur_dist: 0.2207369 
#> target_dist - cur_dist: 0.1707369 
#> target_dist - cur_dist: 0.1207369 
#> target_dist - cur_dist: 0.07073689 
#> target_dist - cur_dist: 0.02073689 
#> generation:  dist =   1.219314 
#> target_dist - cur_dist: 1.219314 
#> target_dist - cur_dist: 1.169314 
#> target_dist - cur_dist: 1.119314 
#> target_dist - cur_dist: 1.069314 
#> target_dist - cur_dist: 1.019314 
#> target_dist - cur_dist: 0.9693137 
#> target_dist - cur_dist: 0.9193137 
#> target_dist - cur_dist: 0.8693137 
#> target_dist - cur_dist: 0.8193137 
#> target_dist - cur_dist: 0.7693137 
#> target_dist - cur_dist: 0.7193137 
#> target_dist - cur_dist: 0.6693137 
#> target_dist - cur_dist: 0.6193137 
#> target_dist - cur_dist: 0.5693137 
#> target_dist - cur_dist: 0.5193137 
#> target_dist - cur_dist: 0.4693137 
#> target_dist - cur_dist: 0.4193137 
#> target_dist - cur_dist: 0.3693137 
#> target_dist - cur_dist: 0.3193137 
#> target_dist - cur_dist: 0.2693137 
#> target_dist - cur_dist: 0.2193137 
#> target_dist - cur_dist: 0.1693137 
#> target_dist - cur_dist: 0.1193137 
#> target_dist - cur_dist: 0.0693137 
#> target_dist - cur_dist: 0.0193137 
#> generation:  dist =   1.290812 
#> target_dist - cur_dist: 1.290812 
#> target_dist - cur_dist: 1.240812 
#> target_dist - cur_dist: 1.190812 
#> target_dist - cur_dist: 1.140812 
#> target_dist - cur_dist: 1.090812 
#> target_dist - cur_dist: 1.040812 
#> target_dist - cur_dist: 0.9908119 
#> target_dist - cur_dist: 0.9408119 
#> target_dist - cur_dist: 0.8908119 
#> target_dist - cur_dist: 0.8408119 
#> target_dist - cur_dist: 0.7908119 
#> target_dist - cur_dist: 0.7408119 
#> target_dist - cur_dist: 0.6908119 
#> target_dist - cur_dist: 0.6408119 
#> target_dist - cur_dist: 0.5908119 
#> target_dist - cur_dist: 0.5408119 
#> target_dist - cur_dist: 0.4908119 
#> target_dist - cur_dist: 0.4408119 
#> target_dist - cur_dist: 0.3908119 
#> target_dist - cur_dist: 0.3408119 
#> target_dist - cur_dist: 0.2908119 
#> target_dist - cur_dist: 0.2408119 
#> target_dist - cur_dist: 0.1908119 
#> target_dist - cur_dist: 0.1408119 
#> target_dist - cur_dist: 0.09081192 
#> target_dist - cur_dist: 0.04081192 
#> generation:  dist =   1.394297 
#> target_dist - cur_dist: 1.394297 
#> target_dist - cur_dist: 1.344297 
#> target_dist - cur_dist: 1.294297 
#> target_dist - cur_dist: 1.244297 
#> target_dist - cur_dist: 1.194297 
#> target_dist - cur_dist: 1.144297 
#> target_dist - cur_dist: 1.094297 
#> target_dist - cur_dist: 1.044297 
#> target_dist - cur_dist: 0.9942971 
#> target_dist - cur_dist: 0.9442971 
#> target_dist - cur_dist: 0.8942971 
#> target_dist - cur_dist: 0.8442971 
#> target_dist - cur_dist: 0.7942971 
#> target_dist - cur_dist: 0.7442971 
#> target_dist - cur_dist: 0.6942971 
#> target_dist - cur_dist: 0.6442971 
#> target_dist - cur_dist: 0.5942971 
#> target_dist - cur_dist: 0.5442971 
#> target_dist - cur_dist: 0.4942971 
#> target_dist - cur_dist: 0.4442971 
#> target_dist - cur_dist: 0.3942971 
#> target_dist - cur_dist: 0.3442971 
#> target_dist - cur_dist: 0.2942971 
#> target_dist - cur_dist: 0.2442971 
#> target_dist - cur_dist: 0.1942971 
#> target_dist - cur_dist: 0.1442971 
#> target_dist - cur_dist: 0.09429712 
#> target_dist - cur_dist: 0.04429712 
#> generation:  dist =   1.011288 
#> target_dist - cur_dist: 1.011288 
#> target_dist - cur_dist: 0.9612877 
#> target_dist - cur_dist: 0.9112877 
#> target_dist - cur_dist: 0.8612877 
#> target_dist - cur_dist: 0.8112877 
#> target_dist - cur_dist: 0.7612877 
#> target_dist - cur_dist: 0.7112877 
#> target_dist - cur_dist: 0.6612877 
#> target_dist - cur_dist: 0.6112877 
#> target_dist - cur_dist: 0.5612877 
#> target_dist - cur_dist: 0.5112877 
#> target_dist - cur_dist: 0.4612877 
#> target_dist - cur_dist: 0.4112877 
#> target_dist - cur_dist: 0.3612877 
#> target_dist - cur_dist: 0.3112877 
#> target_dist - cur_dist: 0.2612877 
#> target_dist - cur_dist: 0.2112877 
#> target_dist - cur_dist: 0.1612877 
#> target_dist - cur_dist: 0.1112877 
#> target_dist - cur_dist: 0.06128767 
#> target_dist - cur_dist: 0.01128767 
#> generation:  dist =   1.291949 
#> target_dist - cur_dist: 1.291949 
#> target_dist - cur_dist: 1.241949 
#> target_dist - cur_dist: 1.191949 
#> target_dist - cur_dist: 1.141949 
#> target_dist - cur_dist: 1.091949 
#> target_dist - cur_dist: 1.041949 
#> target_dist - cur_dist: 0.9919487 
#> target_dist - cur_dist: 0.9419487 
#> target_dist - cur_dist: 0.8919487 
#> target_dist - cur_dist: 0.8419487 
#> target_dist - cur_dist: 0.7919487 
#> target_dist - cur_dist: 0.7419487 
#> target_dist - cur_dist: 0.6919487 
#> target_dist - cur_dist: 0.6419487 
#> target_dist - cur_dist: 0.5919487 
#> target_dist - cur_dist: 0.5419487 
#> target_dist - cur_dist: 0.4919487 
#> target_dist - cur_dist: 0.4419487 
#> target_dist - cur_dist: 0.3919487 
#> target_dist - cur_dist: 0.3419487 
#> target_dist - cur_dist: 0.2919487 
#> target_dist - cur_dist: 0.2419487 
#> target_dist - cur_dist: 0.1919487 
#> target_dist - cur_dist: 0.1419487 
#> target_dist - cur_dist: 0.09194875 
#> target_dist - cur_dist: 0.04194875 

#> [1]   6   1 235

dim(interpolate(t1, 0.1))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.299216 
#> target_dist - cur_dist: 1.299216 
#> target_dist - cur_dist: 1.199216 
#> target_dist - cur_dist: 1.099216 
#> target_dist - cur_dist: 0.9992157 
#> target_dist - cur_dist: 0.8992157 
#> target_dist - cur_dist: 0.7992157 
#> target_dist - cur_dist: 0.6992157 
#> target_dist - cur_dist: 0.5992157 
#> target_dist - cur_dist: 0.4992157 
#> target_dist - cur_dist: 0.3992157 
#> target_dist - cur_dist: 0.2992157 
#> target_dist - cur_dist: 0.1992157 
#> target_dist - cur_dist: 0.09921574 
#> generation:  dist =   1.348672 
#> target_dist - cur_dist: 1.348672 
#> target_dist - cur_dist: 1.248672 
#> target_dist - cur_dist: 1.148672 
#> target_dist - cur_dist: 1.048672 
#> target_dist - cur_dist: 0.9486724 
#> target_dist - cur_dist: 0.8486724 
#> target_dist - cur_dist: 0.7486724 
#> target_dist - cur_dist: 0.6486724 
#> target_dist - cur_dist: 0.5486724 
#> target_dist - cur_dist: 0.4486724 
#> target_dist - cur_dist: 0.3486724 
#> target_dist - cur_dist: 0.2486724 
#> target_dist - cur_dist: 0.1486724 
#> target_dist - cur_dist: 0.04867236 
#> generation:  dist =   1.344725 
#> target_dist - cur_dist: 1.344725 
#> target_dist - cur_dist: 1.244725 
#> target_dist - cur_dist: 1.144725 
#> target_dist - cur_dist: 1.044725 
#> target_dist - cur_dist: 0.9447249 
#> target_dist - cur_dist: 0.8447249 
#> target_dist - cur_dist: 0.7447249 
#> target_dist - cur_dist: 0.6447249 
#> target_dist - cur_dist: 0.5447249 
#> target_dist - cur_dist: 0.4447249 
#> target_dist - cur_dist: 0.3447249 
#> target_dist - cur_dist: 0.2447249 
#> target_dist - cur_dist: 0.1447249 
#> target_dist - cur_dist: 0.04472494 
#> generation:  dist =   1.370737 
#> target_dist - cur_dist: 1.370737 
#> target_dist - cur_dist: 1.270737 
#> target_dist - cur_dist: 1.170737 
#> target_dist - cur_dist: 1.070737 
#> target_dist - cur_dist: 0.9707369 
#> target_dist - cur_dist: 0.8707369 
#> target_dist - cur_dist: 0.7707369 
#> target_dist - cur_dist: 0.6707369 
#> target_dist - cur_dist: 0.5707369 
#> target_dist - cur_dist: 0.4707369 
#> target_dist - cur_dist: 0.3707369 
#> target_dist - cur_dist: 0.2707369 
#> target_dist - cur_dist: 0.1707369 
#> target_dist - cur_dist: 0.07073689 
#> generation:  dist =   1.219314 
#> target_dist - cur_dist: 1.219314 
#> target_dist - cur_dist: 1.119314 
#> target_dist - cur_dist: 1.019314 
#> target_dist - cur_dist: 0.9193137 
#> target_dist - cur_dist: 0.8193137 
#> target_dist - cur_dist: 0.7193137 
#> target_dist - cur_dist: 0.6193137 
#> target_dist - cur_dist: 0.5193137 
#> target_dist - cur_dist: 0.4193137 
#> target_dist - cur_dist: 0.3193137 
#> target_dist - cur_dist: 0.2193137 
#> target_dist - cur_dist: 0.1193137 
#> target_dist - cur_dist: 0.0193137 
#> generation:  dist =   1.290812 
#> target_dist - cur_dist: 1.290812 
#> target_dist - cur_dist: 1.190812 
#> target_dist - cur_dist: 1.090812 
#> target_dist - cur_dist: 0.9908119 
#> target_dist - cur_dist: 0.8908119 
#> target_dist - cur_dist: 0.7908119 
#> target_dist - cur_dist: 0.6908119 
#> target_dist - cur_dist: 0.5908119 
#> target_dist - cur_dist: 0.4908119 
#> target_dist - cur_dist: 0.3908119 
#> target_dist - cur_dist: 0.2908119 
#> target_dist - cur_dist: 0.1908119 
#> target_dist - cur_dist: 0.09081192 
#> generation:  dist =   1.394297 
#> target_dist - cur_dist: 1.394297 
#> target_dist - cur_dist: 1.294297 
#> target_dist - cur_dist: 1.194297 
#> target_dist - cur_dist: 1.094297 
#> target_dist - cur_dist: 0.9942971 
#> target_dist - cur_dist: 0.8942971 
#> target_dist - cur_dist: 0.7942971 
#> target_dist - cur_dist: 0.6942971 
#> target_dist - cur_dist: 0.5942971 
#> target_dist - cur_dist: 0.4942971 
#> target_dist - cur_dist: 0.3942971 
#> target_dist - cur_dist: 0.2942971 
#> target_dist - cur_dist: 0.1942971 
#> target_dist - cur_dist: 0.09429712 
#> generation:  dist =   1.011288 
#> target_dist - cur_dist: 1.011288 
#> target_dist - cur_dist: 0.9112877 
#> target_dist - cur_dist: 0.8112877 
#> target_dist - cur_dist: 0.7112877 
#> target_dist - cur_dist: 0.6112877 
#> target_dist - cur_dist: 0.5112877 
#> target_dist - cur_dist: 0.4112877 
#> target_dist - cur_dist: 0.3112877 
#> target_dist - cur_dist: 0.2112877 
#> target_dist - cur_dist: 0.1112877 
#> target_dist - cur_dist: 0.01128767 
#> generation:  dist =   1.291949 
#> target_dist - cur_dist: 1.291949 
#> target_dist - cur_dist: 1.191949 
#> target_dist - cur_dist: 1.091949 
#> target_dist - cur_dist: 0.9919487 
#> target_dist - cur_dist: 0.8919487 
#> target_dist - cur_dist: 0.7919487 
#> target_dist - cur_dist: 0.6919487 
#> target_dist - cur_dist: 0.5919487 
#> target_dist - cur_dist: 0.4919487 
#> target_dist - cur_dist: 0.3919487 
#> target_dist - cur_dist: 0.2919487 
#> target_dist - cur_dist: 0.1919487 
#> target_dist - cur_dist: 0.09194875 

#> [1]   6   1 120