Render frames of animation to disk

Render frames of animation to disk

render(
  data,
  tour_path,
  display,
  dev,
  ...,
  apf = 1/10,
  frames = 50,
  rescale = TRUE,
  sphere = FALSE,
  start = NULL
)

Arguments

data
matrix, or data frame containing numeric columns

tour_path
tour path generator

display
the method used to render the projected data, e.g. display_xy, display_pcp

dev
name of output device to use (e.g. png, pdf)

...
other options passed to output device

apf
angle (in radians) per frame

frames
number of frames in output

rescale
if true, rescale all variables to range [0,1]

sphere
if true, sphere all variables

start
starting projection. If NULL, uses path default.

References

Hadley Wickham, Dianne Cook, Heike Hofmann, Andreas Buja (2011). tourr: An R Package for Exploring Multivariate Data with Projections. Journal of Statistical Software, 40(2), 1-18. http://www.jstatsoft.org/v40/i02/.

Examples

tmp_path <- tempdir()
render(flea[, 1:4], grand_tour(), display_xy(), "pdf", frames = 10,
  file.path(tmp_path, "test.pdf"))

#> Converting input data to the required matrix format.

#> Using half_range 0.79

render(flea[, 1:4], grand_tour(), display_xy(), "png", frames = 10,
  file.path(tmp_path, "test-%03d.png"))

#> Converting input data to the required matrix format.

#> Using half_range 0.79