Section pursuit index.

Calculates a section pursuit index that compares the distribution inside and outside a slice.

Usage

slice_index(
  breaks_x,
  breaks_y,
  eps,
  bintype = "polar",
  power = 1,
  flip = 1,
  reweight = FALSE,
  p = 4
)

Arguments

breaks_x

binning on the first variable (x or radius).

breaks_y

binning on the second variable (y or angle).

eps

cutoff values to suppress summing up small differences. Vector with one entry for each bin, can be estimated using estimate_eps.

bintype

select polar (default) or square binning.

power

exponent q used in the index compuatation.

flip

sign of the index computation, select +1 when searching for low densities and -1 when searching for high densities.

reweight

if TRUE will reweight according to the expected distribution in a uniform hypersphere (default is FALSE).

p

number of variables in the data (needed for accurate reweighting, default is 4).