Display tour path with trails

Animate a 2D tour path with a point trails

display_trails(
  center = TRUE,
  axes = "center",
  half_range = NULL,
  col = "black",
  pch = 20,
  cex = 1,
  past = 3,
  ...
)

animate_trails(data, tour_path = grand_tour(), ...)

Arguments

center
if TRUE, centers projected data to (0,0). This pins the center of data cloud and make it easier to focus on the changing shape rather than position.

axes
position of the axes: center, bottomleft or off

half_range
half range to use when calculating limits of projected. If not set, defaults to maximum distance from origin to each row of data.

col
color to be plotted. Defaults to "black"

pch
shape of the point to be plotted. Defaults to 20.

cex
magnification of plotting text relative to default. Defaults to 1.

past
draw line between current projection and projection past steps ago

...
other arguments passed on to animate and display_xy

data
matrix, or data frame containing numeric columns

tour_path
tour path generator, defaults to 2d grand tour