Freeze and thaw matrices

Some terminology: * frozen variables: the variables that have fixed values * warm variables: the remaining variables that vary freely

freeze(input, frozen)

thaw(input, frozen)

Details

A frozen matrix specifies which variables to fix in a projection matrix. Warm variables should be missing (NA) while frozen variables should be set to their fixed values.

Examples

frozen <- matrix(NA, nrow = 4, ncol = 2)
frozen[3, ] <- .5

input <- basis_random(4, 2)
freeze(input, frozen)

#>           [,1]       [,2]
#> [1,] 0.1230340 -0.5907730
#> [2,] 0.7706156  0.5368136
#> [3,] 0.0000000  0.0000000
#> [4,] 0.6130703 -0.4842675

thaw(input, frozen)

#>           [,1]       [,2]
#> [1,] 0.1065505 -0.5116244
#> [2,] 0.6673727  0.4648942
#> [3,] 0.5000000  0.5000000
#> [4,] 0.5309344 -0.4193879

freeze(basis_random(4, 2), frozen)

#>            [,1]        [,2]
#> [1,]  0.2259503 -0.62654125
#> [2,]  0.6156134  0.28478699
#> [3,]  0.0000000  0.00000000
#> [4,] -0.7548885  0.05475967