Interpolate geodesically between bases.

This function takes a set of bases and produces a tour by geodesically interpolating between each basis

interpolate(basis_set, angle = 0.05, cycle = FALSE)

Arguments

basis_set
input basis set

angle
target distance (in radians) between bases

cycle
For planned_tour cycle through continuously (TRUE) or stop after first pass (FALSE)

Examples

t1 <- save_history(flea[, 1:6], grand_tour(1), max = 10)

#> Converting input data to the required matrix format.

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.411559 
#> target_dist - cur_dist: 1.411559 
#> generation:  dist =   1.358309 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.374828 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.218371 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.393842 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.151928 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.144572 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.372434 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.090901 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.392163 
#> target_dist - cur_dist: 0 
#> generation:  dist =   1.3389 

dim(t1)

#> [1]  6  1 10

dim(interpolate(t1, 0.01))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.374828 
#> target_dist - cur_dist: 1.374828 
#> target_dist - cur_dist: 1.364828 
#> target_dist - cur_dist: 1.354828 
#> target_dist - cur_dist: 1.344828 
#> target_dist - cur_dist: 1.334828 
#> target_dist - cur_dist: 1.324828 
#> target_dist - cur_dist: 1.314828 
#> target_dist - cur_dist: 1.304828 
#> target_dist - cur_dist: 1.294828 
#> target_dist - cur_dist: 1.284828 
#> target_dist - cur_dist: 1.274828 
#> target_dist - cur_dist: 1.264828 
#> target_dist - cur_dist: 1.254828 
#> target_dist - cur_dist: 1.244828 
#> target_dist - cur_dist: 1.234828 
#> target_dist - cur_dist: 1.224828 
#> target_dist - cur_dist: 1.214828 
#> target_dist - cur_dist: 1.204828 
#> target_dist - cur_dist: 1.194828 
#> target_dist - cur_dist: 1.184828 
#> target_dist - cur_dist: 1.174828 
#> target_dist - cur_dist: 1.164828 
#> target_dist - cur_dist: 1.154828 
#> target_dist - cur_dist: 1.144828 
#> target_dist - cur_dist: 1.134828 
#> target_dist - cur_dist: 1.124828 
#> target_dist - cur_dist: 1.114828 
#> target_dist - cur_dist: 1.104828 
#> target_dist - cur_dist: 1.094828 
#> target_dist - cur_dist: 1.084828 
#> target_dist - cur_dist: 1.074828 
#> target_dist - cur_dist: 1.064828 
#> target_dist - cur_dist: 1.054828 
#> target_dist - cur_dist: 1.044828 
#> target_dist - cur_dist: 1.034828 
#> target_dist - cur_dist: 1.024828 
#> target_dist - cur_dist: 1.014828 
#> target_dist - cur_dist: 1.004828 
#> target_dist - cur_dist: 0.9948283 
#> target_dist - cur_dist: 0.9848283 
#> target_dist - cur_dist: 0.9748283 
#> target_dist - cur_dist: 0.9648283 
#> target_dist - cur_dist: 0.9548283 
#> target_dist - cur_dist: 0.9448283 
#> target_dist - cur_dist: 0.9348283 
#> target_dist - cur_dist: 0.9248283 
#> target_dist - cur_dist: 0.9148283 
#> target_dist - cur_dist: 0.9048283 
#> target_dist - cur_dist: 0.8948283 
#> target_dist - cur_dist: 0.8848283 
#> target_dist - cur_dist: 0.8748283 
#> target_dist - cur_dist: 0.8648283 
#> target_dist - cur_dist: 0.8548283 
#> target_dist - cur_dist: 0.8448283 
#> target_dist - cur_dist: 0.8348283 
#> target_dist - cur_dist: 0.8248283 
#> target_dist - cur_dist: 0.8148283 
#> target_dist - cur_dist: 0.8048283 
#> target_dist - cur_dist: 0.7948283 
#> target_dist - cur_dist: 0.7848283 
#> target_dist - cur_dist: 0.7748283 
#> target_dist - cur_dist: 0.7648283 
#> target_dist - cur_dist: 0.7548283 
#> target_dist - cur_dist: 0.7448283 
#> target_dist - cur_dist: 0.7348283 
#> target_dist - cur_dist: 0.7248283 
#> target_dist - cur_dist: 0.7148283 
#> target_dist - cur_dist: 0.7048283 
#> target_dist - cur_dist: 0.6948283 
#> target_dist - cur_dist: 0.6848283 
#> target_dist - cur_dist: 0.6748283 
#> target_dist - cur_dist: 0.6648283 
#> target_dist - cur_dist: 0.6548283 
#> target_dist - cur_dist: 0.6448283 
#> target_dist - cur_dist: 0.6348283 
#> target_dist - cur_dist: 0.6248283 
#> target_dist - cur_dist: 0.6148283 
#> target_dist - cur_dist: 0.6048283 
#> target_dist - cur_dist: 0.5948283 
#> target_dist - cur_dist: 0.5848283 
#> target_dist - cur_dist: 0.5748283 
#> target_dist - cur_dist: 0.5648283 
#> target_dist - cur_dist: 0.5548283 
#> target_dist - cur_dist: 0.5448283 
#> target_dist - cur_dist: 0.5348283 
#> target_dist - cur_dist: 0.5248283 
#> target_dist - cur_dist: 0.5148283 
#> target_dist - cur_dist: 0.5048283 
#> target_dist - cur_dist: 0.4948283 
#> target_dist - cur_dist: 0.4848283 
#> target_dist - cur_dist: 0.4748283 
#> target_dist - cur_dist: 0.4648283 
#> target_dist - cur_dist: 0.4548283 
#> target_dist - cur_dist: 0.4448283 
#> target_dist - cur_dist: 0.4348283 
#> target_dist - cur_dist: 0.4248283 
#> target_dist - cur_dist: 0.4148283 
#> target_dist - cur_dist: 0.4048283 
#> target_dist - cur_dist: 0.3948283 
#> target_dist - cur_dist: 0.3848283 
#> target_dist - cur_dist: 0.3748283 
#> target_dist - cur_dist: 0.3648283 
#> target_dist - cur_dist: 0.3548283 
#> target_dist - cur_dist: 0.3448283 
#> target_dist - cur_dist: 0.3348283 
#> target_dist - cur_dist: 0.3248283 
#> target_dist - cur_dist: 0.3148283 
#> target_dist - cur_dist: 0.3048283 
#> target_dist - cur_dist: 0.2948283 
#> target_dist - cur_dist: 0.2848283 
#> target_dist - cur_dist: 0.2748283 
#> target_dist - cur_dist: 0.2648283 
#> target_dist - cur_dist: 0.2548283 
#> target_dist - cur_dist: 0.2448283 
#> target_dist - cur_dist: 0.2348283 
#> target_dist - cur_dist: 0.2248283 
#> target_dist - cur_dist: 0.2148283 
#> target_dist - cur_dist: 0.2048283 
#> target_dist - cur_dist: 0.1948283 
#> target_dist - cur_dist: 0.1848283 
#> target_dist - cur_dist: 0.1748283 
#> target_dist - cur_dist: 0.1648283 
#> target_dist - cur_dist: 0.1548283 
#> target_dist - cur_dist: 0.1448283 
#> target_dist - cur_dist: 0.1348283 
#> target_dist - cur_dist: 0.1248283 
#> target_dist - cur_dist: 0.1148283 
#> target_dist - cur_dist: 0.1048283 
#> target_dist - cur_dist: 0.09482834 
#> target_dist - cur_dist: 0.08482834 
#> target_dist - cur_dist: 0.07482834 
#> target_dist - cur_dist: 0.06482834 
#> target_dist - cur_dist: 0.05482834 
#> target_dist - cur_dist: 0.04482834 
#> target_dist - cur_dist: 0.03482834 
#> target_dist - cur_dist: 0.02482834 
#> target_dist - cur_dist: 0.01482834 
#> target_dist - cur_dist: 0.004828339 
#> generation:  dist =   1.218371 
#> target_dist - cur_dist: 1.218371 
#> target_dist - cur_dist: 1.208371 
#> target_dist - cur_dist: 1.198371 
#> target_dist - cur_dist: 1.188371 
#> target_dist - cur_dist: 1.178371 
#> target_dist - cur_dist: 1.168371 
#> target_dist - cur_dist: 1.158371 
#> target_dist - cur_dist: 1.148371 
#> target_dist - cur_dist: 1.138371 
#> target_dist - cur_dist: 1.128371 
#> target_dist - cur_dist: 1.118371 
#> target_dist - cur_dist: 1.108371 
#> target_dist - cur_dist: 1.098371 
#> target_dist - cur_dist: 1.088371 
#> target_dist - cur_dist: 1.078371 
#> target_dist - cur_dist: 1.068371 
#> target_dist - cur_dist: 1.058371 
#> target_dist - cur_dist: 1.048371 
#> target_dist - cur_dist: 1.038371 
#> target_dist - cur_dist: 1.028371 
#> target_dist - cur_dist: 1.018371 
#> target_dist - cur_dist: 1.008371 
#> target_dist - cur_dist: 0.9983711 
#> target_dist - cur_dist: 0.9883711 
#> target_dist - cur_dist: 0.9783711 
#> target_dist - cur_dist: 0.9683711 
#> target_dist - cur_dist: 0.9583711 
#> target_dist - cur_dist: 0.9483711 
#> target_dist - cur_dist: 0.9383711 
#> target_dist - cur_dist: 0.9283711 
#> target_dist - cur_dist: 0.9183711 
#> target_dist - cur_dist: 0.9083711 
#> target_dist - cur_dist: 0.8983711 
#> target_dist - cur_dist: 0.8883711 
#> target_dist - cur_dist: 0.8783711 
#> target_dist - cur_dist: 0.8683711 
#> target_dist - cur_dist: 0.8583711 
#> target_dist - cur_dist: 0.8483711 
#> target_dist - cur_dist: 0.8383711 
#> target_dist - cur_dist: 0.8283711 
#> target_dist - cur_dist: 0.8183711 
#> target_dist - cur_dist: 0.8083711 
#> target_dist - cur_dist: 0.7983711 
#> target_dist - cur_dist: 0.7883711 
#> target_dist - cur_dist: 0.7783711 
#> target_dist - cur_dist: 0.7683711 
#> target_dist - cur_dist: 0.7583711 
#> target_dist - cur_dist: 0.7483711 
#> target_dist - cur_dist: 0.7383711 
#> target_dist - cur_dist: 0.7283711 
#> target_dist - cur_dist: 0.7183711 
#> target_dist - cur_dist: 0.7083711 
#> target_dist - cur_dist: 0.6983711 
#> target_dist - cur_dist: 0.6883711 
#> target_dist - cur_dist: 0.6783711 
#> target_dist - cur_dist: 0.6683711 
#> target_dist - cur_dist: 0.6583711 
#> target_dist - cur_dist: 0.6483711 
#> target_dist - cur_dist: 0.6383711 
#> target_dist - cur_dist: 0.6283711 
#> target_dist - cur_dist: 0.6183711 
#> target_dist - cur_dist: 0.6083711 
#> target_dist - cur_dist: 0.5983711 
#> target_dist - cur_dist: 0.5883711 
#> target_dist - cur_dist: 0.5783711 
#> target_dist - cur_dist: 0.5683711 
#> target_dist - cur_dist: 0.5583711 
#> target_dist - cur_dist: 0.5483711 
#> target_dist - cur_dist: 0.5383711 
#> target_dist - cur_dist: 0.5283711 
#> target_dist - cur_dist: 0.5183711 
#> target_dist - cur_dist: 0.5083711 
#> target_dist - cur_dist: 0.4983711 
#> target_dist - cur_dist: 0.4883711 
#> target_dist - cur_dist: 0.4783711 
#> target_dist - cur_dist: 0.4683711 
#> target_dist - cur_dist: 0.4583711 
#> target_dist - cur_dist: 0.4483711 
#> target_dist - cur_dist: 0.4383711 
#> target_dist - cur_dist: 0.4283711 
#> target_dist - cur_dist: 0.4183711 
#> target_dist - cur_dist: 0.4083711 
#> target_dist - cur_dist: 0.3983711 
#> target_dist - cur_dist: 0.3883711 
#> target_dist - cur_dist: 0.3783711 
#> target_dist - cur_dist: 0.3683711 
#> target_dist - cur_dist: 0.3583711 
#> target_dist - cur_dist: 0.3483711 
#> target_dist - cur_dist: 0.3383711 
#> target_dist - cur_dist: 0.3283711 
#> target_dist - cur_dist: 0.3183711 
#> target_dist - cur_dist: 0.3083711 
#> target_dist - cur_dist: 0.2983711 
#> target_dist - cur_dist: 0.2883711 
#> target_dist - cur_dist: 0.2783711 
#> target_dist - cur_dist: 0.2683711 
#> target_dist - cur_dist: 0.2583711 
#> target_dist - cur_dist: 0.2483711 
#> target_dist - cur_dist: 0.2383711 
#> target_dist - cur_dist: 0.2283711 
#> target_dist - cur_dist: 0.2183711 
#> target_dist - cur_dist: 0.2083711 
#> target_dist - cur_dist: 0.1983711 
#> target_dist - cur_dist: 0.1883711 
#> target_dist - cur_dist: 0.1783711 
#> target_dist - cur_dist: 0.1683711 
#> target_dist - cur_dist: 0.1583711 
#> target_dist - cur_dist: 0.1483711 
#> target_dist - cur_dist: 0.1383711 
#> target_dist - cur_dist: 0.1283711 
#> target_dist - cur_dist: 0.1183711 
#> target_dist - cur_dist: 0.1083711 
#> target_dist - cur_dist: 0.09837115 
#> target_dist - cur_dist: 0.08837115 
#> target_dist - cur_dist: 0.07837115 
#> target_dist - cur_dist: 0.06837115 
#> target_dist - cur_dist: 0.05837115 
#> target_dist - cur_dist: 0.04837115 
#> target_dist - cur_dist: 0.03837115 
#> target_dist - cur_dist: 0.02837115 
#> target_dist - cur_dist: 0.01837115 
#> target_dist - cur_dist: 0.008371147 
#> generation:  dist =   1.393842 
#> target_dist - cur_dist: 1.393842 
#> target_dist - cur_dist: 1.383842 
#> target_dist - cur_dist: 1.373842 
#> target_dist - cur_dist: 1.363842 
#> target_dist - cur_dist: 1.353842 
#> target_dist - cur_dist: 1.343842 
#> target_dist - cur_dist: 1.333842 
#> target_dist - cur_dist: 1.323842 
#> target_dist - cur_dist: 1.313842 
#> target_dist - cur_dist: 1.303842 
#> target_dist - cur_dist: 1.293842 
#> target_dist - cur_dist: 1.283842 
#> target_dist - cur_dist: 1.273842 
#> target_dist - cur_dist: 1.263842 
#> target_dist - cur_dist: 1.253842 
#> target_dist - cur_dist: 1.243842 
#> target_dist - cur_dist: 1.233842 
#> target_dist - cur_dist: 1.223842 
#> target_dist - cur_dist: 1.213842 
#> target_dist - cur_dist: 1.203842 
#> target_dist - cur_dist: 1.193842 
#> target_dist - cur_dist: 1.183842 
#> target_dist - cur_dist: 1.173842 
#> target_dist - cur_dist: 1.163842 
#> target_dist - cur_dist: 1.153842 
#> target_dist - cur_dist: 1.143842 
#> target_dist - cur_dist: 1.133842 
#> target_dist - cur_dist: 1.123842 
#> target_dist - cur_dist: 1.113842 
#> target_dist - cur_dist: 1.103842 
#> target_dist - cur_dist: 1.093842 
#> target_dist - cur_dist: 1.083842 
#> target_dist - cur_dist: 1.073842 
#> target_dist - cur_dist: 1.063842 
#> target_dist - cur_dist: 1.053842 
#> target_dist - cur_dist: 1.043842 
#> target_dist - cur_dist: 1.033842 
#> target_dist - cur_dist: 1.023842 
#> target_dist - cur_dist: 1.013842 
#> target_dist - cur_dist: 1.003842 
#> target_dist - cur_dist: 0.9938419 
#> target_dist - cur_dist: 0.9838419 
#> target_dist - cur_dist: 0.9738419 
#> target_dist - cur_dist: 0.9638419 
#> target_dist - cur_dist: 0.9538419 
#> target_dist - cur_dist: 0.9438419 
#> target_dist - cur_dist: 0.9338419 
#> target_dist - cur_dist: 0.9238419 
#> target_dist - cur_dist: 0.9138419 
#> target_dist - cur_dist: 0.9038419 
#> target_dist - cur_dist: 0.8938419 
#> target_dist - cur_dist: 0.8838419 
#> target_dist - cur_dist: 0.8738419 
#> target_dist - cur_dist: 0.8638419 
#> target_dist - cur_dist: 0.8538419 
#> target_dist - cur_dist: 0.8438419 
#> target_dist - cur_dist: 0.8338419 
#> target_dist - cur_dist: 0.8238419 
#> target_dist - cur_dist: 0.8138419 
#> target_dist - cur_dist: 0.8038419 
#> target_dist - cur_dist: 0.7938419 
#> target_dist - cur_dist: 0.7838419 
#> target_dist - cur_dist: 0.7738419 
#> target_dist - cur_dist: 0.7638419 
#> target_dist - cur_dist: 0.7538419 
#> target_dist - cur_dist: 0.7438419 
#> target_dist - cur_dist: 0.7338419 
#> target_dist - cur_dist: 0.7238419 
#> target_dist - cur_dist: 0.7138419 
#> target_dist - cur_dist: 0.7038419 
#> target_dist - cur_dist: 0.6938419 
#> target_dist - cur_dist: 0.6838419 
#> target_dist - cur_dist: 0.6738419 
#> target_dist - cur_dist: 0.6638419 
#> target_dist - cur_dist: 0.6538419 
#> target_dist - cur_dist: 0.6438419 
#> target_dist - cur_dist: 0.6338419 
#> target_dist - cur_dist: 0.6238419 
#> target_dist - cur_dist: 0.6138419 
#> target_dist - cur_dist: 0.6038419 
#> target_dist - cur_dist: 0.5938419 
#> target_dist - cur_dist: 0.5838419 
#> target_dist - cur_dist: 0.5738419 
#> target_dist - cur_dist: 0.5638419 
#> target_dist - cur_dist: 0.5538419 
#> target_dist - cur_dist: 0.5438419 
#> target_dist - cur_dist: 0.5338419 
#> target_dist - cur_dist: 0.5238419 
#> target_dist - cur_dist: 0.5138419 
#> target_dist - cur_dist: 0.5038419 
#> target_dist - cur_dist: 0.4938419 
#> target_dist - cur_dist: 0.4838419 
#> target_dist - cur_dist: 0.4738419 
#> target_dist - cur_dist: 0.4638419 
#> target_dist - cur_dist: 0.4538419 
#> target_dist - cur_dist: 0.4438419 
#> target_dist - cur_dist: 0.4338419 
#> target_dist - cur_dist: 0.4238419 
#> target_dist - cur_dist: 0.4138419 
#> target_dist - cur_dist: 0.4038419 
#> target_dist - cur_dist: 0.3938419 
#> target_dist - cur_dist: 0.3838419 
#> target_dist - cur_dist: 0.3738419 
#> target_dist - cur_dist: 0.3638419 
#> target_dist - cur_dist: 0.3538419 
#> target_dist - cur_dist: 0.3438419 
#> target_dist - cur_dist: 0.3338419 
#> target_dist - cur_dist: 0.3238419 
#> target_dist - cur_dist: 0.3138419 
#> target_dist - cur_dist: 0.3038419 
#> target_dist - cur_dist: 0.2938419 
#> target_dist - cur_dist: 0.2838419 
#> target_dist - cur_dist: 0.2738419 
#> target_dist - cur_dist: 0.2638419 
#> target_dist - cur_dist: 0.2538419 
#> target_dist - cur_dist: 0.2438419 
#> target_dist - cur_dist: 0.2338419 
#> target_dist - cur_dist: 0.2238419 
#> target_dist - cur_dist: 0.2138419 
#> target_dist - cur_dist: 0.2038419 
#> target_dist - cur_dist: 0.1938419 
#> target_dist - cur_dist: 0.1838419 
#> target_dist - cur_dist: 0.1738419 
#> target_dist - cur_dist: 0.1638419 
#> target_dist - cur_dist: 0.1538419 
#> target_dist - cur_dist: 0.1438419 
#> target_dist - cur_dist: 0.1338419 
#> target_dist - cur_dist: 0.1238419 
#> target_dist - cur_dist: 0.1138419 
#> target_dist - cur_dist: 0.1038419 
#> target_dist - cur_dist: 0.0938419 
#> target_dist - cur_dist: 0.0838419 
#> target_dist - cur_dist: 0.0738419 
#> target_dist - cur_dist: 0.0638419 
#> target_dist - cur_dist: 0.0538419 
#> target_dist - cur_dist: 0.0438419 
#> target_dist - cur_dist: 0.0338419 
#> target_dist - cur_dist: 0.0238419 
#> target_dist - cur_dist: 0.0138419 
#> target_dist - cur_dist: 0.0038419 
#> generation:  dist =   1.151928 
#> target_dist - cur_dist: 1.151928 
#> target_dist - cur_dist: 1.141928 
#> target_dist - cur_dist: 1.131928 
#> target_dist - cur_dist: 1.121928 
#> target_dist - cur_dist: 1.111928 
#> target_dist - cur_dist: 1.101928 
#> target_dist - cur_dist: 1.091928 
#> target_dist - cur_dist: 1.081928 
#> target_dist - cur_dist: 1.071928 
#> target_dist - cur_dist: 1.061928 
#> target_dist - cur_dist: 1.051928 
#> target_dist - cur_dist: 1.041928 
#> target_dist - cur_dist: 1.031928 
#> target_dist - cur_dist: 1.021928 
#> target_dist - cur_dist: 1.011928 
#> target_dist - cur_dist: 1.001928 
#> target_dist - cur_dist: 0.9919276 
#> target_dist - cur_dist: 0.9819276 
#> target_dist - cur_dist: 0.9719276 
#> target_dist - cur_dist: 0.9619276 
#> target_dist - cur_dist: 0.9519276 
#> target_dist - cur_dist: 0.9419276 
#> target_dist - cur_dist: 0.9319276 
#> target_dist - cur_dist: 0.9219276 
#> target_dist - cur_dist: 0.9119276 
#> target_dist - cur_dist: 0.9019276 
#> target_dist - cur_dist: 0.8919276 
#> target_dist - cur_dist: 0.8819276 
#> target_dist - cur_dist: 0.8719276 
#> target_dist - cur_dist: 0.8619276 
#> target_dist - cur_dist: 0.8519276 
#> target_dist - cur_dist: 0.8419276 
#> target_dist - cur_dist: 0.8319276 
#> target_dist - cur_dist: 0.8219276 
#> target_dist - cur_dist: 0.8119276 
#> target_dist - cur_dist: 0.8019276 
#> target_dist - cur_dist: 0.7919276 
#> target_dist - cur_dist: 0.7819276 
#> target_dist - cur_dist: 0.7719276 
#> target_dist - cur_dist: 0.7619276 
#> target_dist - cur_dist: 0.7519276 
#> target_dist - cur_dist: 0.7419276 
#> target_dist - cur_dist: 0.7319276 
#> target_dist - cur_dist: 0.7219276 
#> target_dist - cur_dist: 0.7119276 
#> target_dist - cur_dist: 0.7019276 
#> target_dist - cur_dist: 0.6919276 
#> target_dist - cur_dist: 0.6819276 
#> target_dist - cur_dist: 0.6719276 
#> target_dist - cur_dist: 0.6619276 
#> target_dist - cur_dist: 0.6519276 
#> target_dist - cur_dist: 0.6419276 
#> target_dist - cur_dist: 0.6319276 
#> target_dist - cur_dist: 0.6219276 
#> target_dist - cur_dist: 0.6119276 
#> target_dist - cur_dist: 0.6019276 
#> target_dist - cur_dist: 0.5919276 
#> target_dist - cur_dist: 0.5819276 
#> target_dist - cur_dist: 0.5719276 
#> target_dist - cur_dist: 0.5619276 
#> target_dist - cur_dist: 0.5519276 
#> target_dist - cur_dist: 0.5419276 
#> target_dist - cur_dist: 0.5319276 
#> target_dist - cur_dist: 0.5219276 
#> target_dist - cur_dist: 0.5119276 
#> target_dist - cur_dist: 0.5019276 
#> target_dist - cur_dist: 0.4919276 
#> target_dist - cur_dist: 0.4819276 
#> target_dist - cur_dist: 0.4719276 
#> target_dist - cur_dist: 0.4619276 
#> target_dist - cur_dist: 0.4519276 
#> target_dist - cur_dist: 0.4419276 
#> target_dist - cur_dist: 0.4319276 
#> target_dist - cur_dist: 0.4219276 
#> target_dist - cur_dist: 0.4119276 
#> target_dist - cur_dist: 0.4019276 
#> target_dist - cur_dist: 0.3919276 
#> target_dist - cur_dist: 0.3819276 
#> target_dist - cur_dist: 0.3719276 
#> target_dist - cur_dist: 0.3619276 
#> target_dist - cur_dist: 0.3519276 
#> target_dist - cur_dist: 0.3419276 
#> target_dist - cur_dist: 0.3319276 
#> target_dist - cur_dist: 0.3219276 
#> target_dist - cur_dist: 0.3119276 
#> target_dist - cur_dist: 0.3019276 
#> target_dist - cur_dist: 0.2919276 
#> target_dist - cur_dist: 0.2819276 
#> target_dist - cur_dist: 0.2719276 
#> target_dist - cur_dist: 0.2619276 
#> target_dist - cur_dist: 0.2519276 
#> target_dist - cur_dist: 0.2419276 
#> target_dist - cur_dist: 0.2319276 
#> target_dist - cur_dist: 0.2219276 
#> target_dist - cur_dist: 0.2119276 
#> target_dist - cur_dist: 0.2019276 
#> target_dist - cur_dist: 0.1919276 
#> target_dist - cur_dist: 0.1819276 
#> target_dist - cur_dist: 0.1719276 
#> target_dist - cur_dist: 0.1619276 
#> target_dist - cur_dist: 0.1519276 
#> target_dist - cur_dist: 0.1419276 
#> target_dist - cur_dist: 0.1319276 
#> target_dist - cur_dist: 0.1219276 
#> target_dist - cur_dist: 0.1119276 
#> target_dist - cur_dist: 0.1019276 
#> target_dist - cur_dist: 0.09192763 
#> target_dist - cur_dist: 0.08192763 
#> target_dist - cur_dist: 0.07192763 
#> target_dist - cur_dist: 0.06192763 
#> target_dist - cur_dist: 0.05192763 
#> target_dist - cur_dist: 0.04192763 
#> target_dist - cur_dist: 0.03192763 
#> target_dist - cur_dist: 0.02192763 
#> target_dist - cur_dist: 0.01192763 
#> target_dist - cur_dist: 0.001927631 
#> generation:  dist =   1.144572 
#> target_dist - cur_dist: 1.144572 
#> target_dist - cur_dist: 1.134572 
#> target_dist - cur_dist: 1.124572 
#> target_dist - cur_dist: 1.114572 
#> target_dist - cur_dist: 1.104572 
#> target_dist - cur_dist: 1.094572 
#> target_dist - cur_dist: 1.084572 
#> target_dist - cur_dist: 1.074572 
#> target_dist - cur_dist: 1.064572 
#> target_dist - cur_dist: 1.054572 
#> target_dist - cur_dist: 1.044572 
#> target_dist - cur_dist: 1.034572 
#> target_dist - cur_dist: 1.024572 
#> target_dist - cur_dist: 1.014572 
#> target_dist - cur_dist: 1.004572 
#> target_dist - cur_dist: 0.994572 
#> target_dist - cur_dist: 0.984572 
#> target_dist - cur_dist: 0.974572 
#> target_dist - cur_dist: 0.964572 
#> target_dist - cur_dist: 0.954572 
#> target_dist - cur_dist: 0.944572 
#> target_dist - cur_dist: 0.934572 
#> target_dist - cur_dist: 0.924572 
#> target_dist - cur_dist: 0.914572 
#> target_dist - cur_dist: 0.904572 
#> target_dist - cur_dist: 0.894572 
#> target_dist - cur_dist: 0.884572 
#> target_dist - cur_dist: 0.874572 
#> target_dist - cur_dist: 0.864572 
#> target_dist - cur_dist: 0.854572 
#> target_dist - cur_dist: 0.844572 
#> target_dist - cur_dist: 0.834572 
#> target_dist - cur_dist: 0.824572 
#> target_dist - cur_dist: 0.814572 
#> target_dist - cur_dist: 0.804572 
#> target_dist - cur_dist: 0.794572 
#> target_dist - cur_dist: 0.784572 
#> target_dist - cur_dist: 0.774572 
#> target_dist - cur_dist: 0.764572 
#> target_dist - cur_dist: 0.754572 
#> target_dist - cur_dist: 0.744572 
#> target_dist - cur_dist: 0.734572 
#> target_dist - cur_dist: 0.724572 
#> target_dist - cur_dist: 0.714572 
#> target_dist - cur_dist: 0.704572 
#> target_dist - cur_dist: 0.694572 
#> target_dist - cur_dist: 0.684572 
#> target_dist - cur_dist: 0.674572 
#> target_dist - cur_dist: 0.664572 
#> target_dist - cur_dist: 0.654572 
#> target_dist - cur_dist: 0.644572 
#> target_dist - cur_dist: 0.634572 
#> target_dist - cur_dist: 0.624572 
#> target_dist - cur_dist: 0.614572 
#> target_dist - cur_dist: 0.604572 
#> target_dist - cur_dist: 0.594572 
#> target_dist - cur_dist: 0.584572 
#> target_dist - cur_dist: 0.574572 
#> target_dist - cur_dist: 0.564572 
#> target_dist - cur_dist: 0.554572 
#> target_dist - cur_dist: 0.544572 
#> target_dist - cur_dist: 0.534572 
#> target_dist - cur_dist: 0.524572 
#> target_dist - cur_dist: 0.514572 
#> target_dist - cur_dist: 0.504572 
#> target_dist - cur_dist: 0.494572 
#> target_dist - cur_dist: 0.484572 
#> target_dist - cur_dist: 0.474572 
#> target_dist - cur_dist: 0.464572 
#> target_dist - cur_dist: 0.454572 
#> target_dist - cur_dist: 0.444572 
#> target_dist - cur_dist: 0.434572 
#> target_dist - cur_dist: 0.424572 
#> target_dist - cur_dist: 0.414572 
#> target_dist - cur_dist: 0.404572 
#> target_dist - cur_dist: 0.394572 
#> target_dist - cur_dist: 0.384572 
#> target_dist - cur_dist: 0.374572 
#> target_dist - cur_dist: 0.364572 
#> target_dist - cur_dist: 0.354572 
#> target_dist - cur_dist: 0.344572 
#> target_dist - cur_dist: 0.334572 
#> target_dist - cur_dist: 0.324572 
#> target_dist - cur_dist: 0.314572 
#> target_dist - cur_dist: 0.304572 
#> target_dist - cur_dist: 0.294572 
#> target_dist - cur_dist: 0.284572 
#> target_dist - cur_dist: 0.274572 
#> target_dist - cur_dist: 0.264572 
#> target_dist - cur_dist: 0.254572 
#> target_dist - cur_dist: 0.244572 
#> target_dist - cur_dist: 0.234572 
#> target_dist - cur_dist: 0.224572 
#> target_dist - cur_dist: 0.214572 
#> target_dist - cur_dist: 0.204572 
#> target_dist - cur_dist: 0.194572 
#> target_dist - cur_dist: 0.184572 
#> target_dist - cur_dist: 0.174572 
#> target_dist - cur_dist: 0.164572 
#> target_dist - cur_dist: 0.154572 
#> target_dist - cur_dist: 0.144572 
#> target_dist - cur_dist: 0.134572 
#> target_dist - cur_dist: 0.124572 
#> target_dist - cur_dist: 0.114572 
#> target_dist - cur_dist: 0.104572 
#> target_dist - cur_dist: 0.09457201 
#> target_dist - cur_dist: 0.08457201 
#> target_dist - cur_dist: 0.07457201 
#> target_dist - cur_dist: 0.06457201 
#> target_dist - cur_dist: 0.05457201 
#> target_dist - cur_dist: 0.04457201 
#> target_dist - cur_dist: 0.03457201 
#> target_dist - cur_dist: 0.02457201 
#> target_dist - cur_dist: 0.01457201 
#> target_dist - cur_dist: 0.004572006 
#> generation:  dist =   1.372434 
#> target_dist - cur_dist: 1.372434 
#> target_dist - cur_dist: 1.362434 
#> target_dist - cur_dist: 1.352434 
#> target_dist - cur_dist: 1.342434 
#> target_dist - cur_dist: 1.332434 
#> target_dist - cur_dist: 1.322434 
#> target_dist - cur_dist: 1.312434 
#> target_dist - cur_dist: 1.302434 
#> target_dist - cur_dist: 1.292434 
#> target_dist - cur_dist: 1.282434 
#> target_dist - cur_dist: 1.272434 
#> target_dist - cur_dist: 1.262434 
#> target_dist - cur_dist: 1.252434 
#> target_dist - cur_dist: 1.242434 
#> target_dist - cur_dist: 1.232434 
#> target_dist - cur_dist: 1.222434 
#> target_dist - cur_dist: 1.212434 
#> target_dist - cur_dist: 1.202434 
#> target_dist - cur_dist: 1.192434 
#> target_dist - cur_dist: 1.182434 
#> target_dist - cur_dist: 1.172434 
#> target_dist - cur_dist: 1.162434 
#> target_dist - cur_dist: 1.152434 
#> target_dist - cur_dist: 1.142434 
#> target_dist - cur_dist: 1.132434 
#> target_dist - cur_dist: 1.122434 
#> target_dist - cur_dist: 1.112434 
#> target_dist - cur_dist: 1.102434 
#> target_dist - cur_dist: 1.092434 
#> target_dist - cur_dist: 1.082434 
#> target_dist - cur_dist: 1.072434 
#> target_dist - cur_dist: 1.062434 
#> target_dist - cur_dist: 1.052434 
#> target_dist - cur_dist: 1.042434 
#> target_dist - cur_dist: 1.032434 
#> target_dist - cur_dist: 1.022434 
#> target_dist - cur_dist: 1.012434 
#> target_dist - cur_dist: 1.002434 
#> target_dist - cur_dist: 0.9924345 
#> target_dist - cur_dist: 0.9824345 
#> target_dist - cur_dist: 0.9724345 
#> target_dist - cur_dist: 0.9624345 
#> target_dist - cur_dist: 0.9524345 
#> target_dist - cur_dist: 0.9424345 
#> target_dist - cur_dist: 0.9324345 
#> target_dist - cur_dist: 0.9224345 
#> target_dist - cur_dist: 0.9124345 
#> target_dist - cur_dist: 0.9024345 
#> target_dist - cur_dist: 0.8924345 
#> target_dist - cur_dist: 0.8824345 
#> target_dist - cur_dist: 0.8724345 
#> target_dist - cur_dist: 0.8624345 
#> target_dist - cur_dist: 0.8524345 
#> target_dist - cur_dist: 0.8424345 
#> target_dist - cur_dist: 0.8324345 
#> target_dist - cur_dist: 0.8224345 
#> target_dist - cur_dist: 0.8124345 
#> target_dist - cur_dist: 0.8024345 
#> target_dist - cur_dist: 0.7924345 
#> target_dist - cur_dist: 0.7824345 
#> target_dist - cur_dist: 0.7724345 
#> target_dist - cur_dist: 0.7624345 
#> target_dist - cur_dist: 0.7524345 
#> target_dist - cur_dist: 0.7424345 
#> target_dist - cur_dist: 0.7324345 
#> target_dist - cur_dist: 0.7224345 
#> target_dist - cur_dist: 0.7124345 
#> target_dist - cur_dist: 0.7024345 
#> target_dist - cur_dist: 0.6924345 
#> target_dist - cur_dist: 0.6824345 
#> target_dist - cur_dist: 0.6724345 
#> target_dist - cur_dist: 0.6624345 
#> target_dist - cur_dist: 0.6524345 
#> target_dist - cur_dist: 0.6424345 
#> target_dist - cur_dist: 0.6324345 
#> target_dist - cur_dist: 0.6224345 
#> target_dist - cur_dist: 0.6124345 
#> target_dist - cur_dist: 0.6024345 
#> target_dist - cur_dist: 0.5924345 
#> target_dist - cur_dist: 0.5824345 
#> target_dist - cur_dist: 0.5724345 
#> target_dist - cur_dist: 0.5624345 
#> target_dist - cur_dist: 0.5524345 
#> target_dist - cur_dist: 0.5424345 
#> target_dist - cur_dist: 0.5324345 
#> target_dist - cur_dist: 0.5224345 
#> target_dist - cur_dist: 0.5124345 
#> target_dist - cur_dist: 0.5024345 
#> target_dist - cur_dist: 0.4924345 
#> target_dist - cur_dist: 0.4824345 
#> target_dist - cur_dist: 0.4724345 
#> target_dist - cur_dist: 0.4624345 
#> target_dist - cur_dist: 0.4524345 
#> target_dist - cur_dist: 0.4424345 
#> target_dist - cur_dist: 0.4324345 
#> target_dist - cur_dist: 0.4224345 
#> target_dist - cur_dist: 0.4124345 
#> target_dist - cur_dist: 0.4024345 
#> target_dist - cur_dist: 0.3924345 
#> target_dist - cur_dist: 0.3824345 
#> target_dist - cur_dist: 0.3724345 
#> target_dist - cur_dist: 0.3624345 
#> target_dist - cur_dist: 0.3524345 
#> target_dist - cur_dist: 0.3424345 
#> target_dist - cur_dist: 0.3324345 
#> target_dist - cur_dist: 0.3224345 
#> target_dist - cur_dist: 0.3124345 
#> target_dist - cur_dist: 0.3024345 
#> target_dist - cur_dist: 0.2924345 
#> target_dist - cur_dist: 0.2824345 
#> target_dist - cur_dist: 0.2724345 
#> target_dist - cur_dist: 0.2624345 
#> target_dist - cur_dist: 0.2524345 
#> target_dist - cur_dist: 0.2424345 
#> target_dist - cur_dist: 0.2324345 
#> target_dist - cur_dist: 0.2224345 
#> target_dist - cur_dist: 0.2124345 
#> target_dist - cur_dist: 0.2024345 
#> target_dist - cur_dist: 0.1924345 
#> target_dist - cur_dist: 0.1824345 
#> target_dist - cur_dist: 0.1724345 
#> target_dist - cur_dist: 0.1624345 
#> target_dist - cur_dist: 0.1524345 
#> target_dist - cur_dist: 0.1424345 
#> target_dist - cur_dist: 0.1324345 
#> target_dist - cur_dist: 0.1224345 
#> target_dist - cur_dist: 0.1124345 
#> target_dist - cur_dist: 0.1024345 
#> target_dist - cur_dist: 0.09243449 
#> target_dist - cur_dist: 0.08243449 
#> target_dist - cur_dist: 0.07243449 
#> target_dist - cur_dist: 0.06243449 
#> target_dist - cur_dist: 0.05243449 
#> target_dist - cur_dist: 0.04243449 
#> target_dist - cur_dist: 0.03243449 
#> target_dist - cur_dist: 0.02243449 
#> target_dist - cur_dist: 0.01243449 
#> target_dist - cur_dist: 0.002434487 
#> generation:  dist =   1.090901 
#> target_dist - cur_dist: 1.090901 
#> target_dist - cur_dist: 1.080901 
#> target_dist - cur_dist: 1.070901 
#> target_dist - cur_dist: 1.060901 
#> target_dist - cur_dist: 1.050901 
#> target_dist - cur_dist: 1.040901 
#> target_dist - cur_dist: 1.030901 
#> target_dist - cur_dist: 1.020901 
#> target_dist - cur_dist: 1.010901 
#> target_dist - cur_dist: 1.000901 
#> target_dist - cur_dist: 0.9909008 
#> target_dist - cur_dist: 0.9809008 
#> target_dist - cur_dist: 0.9709008 
#> target_dist - cur_dist: 0.9609008 
#> target_dist - cur_dist: 0.9509008 
#> target_dist - cur_dist: 0.9409008 
#> target_dist - cur_dist: 0.9309008 
#> target_dist - cur_dist: 0.9209008 
#> target_dist - cur_dist: 0.9109008 
#> target_dist - cur_dist: 0.9009008 
#> target_dist - cur_dist: 0.8909008 
#> target_dist - cur_dist: 0.8809008 
#> target_dist - cur_dist: 0.8709008 
#> target_dist - cur_dist: 0.8609008 
#> target_dist - cur_dist: 0.8509008 
#> target_dist - cur_dist: 0.8409008 
#> target_dist - cur_dist: 0.8309008 
#> target_dist - cur_dist: 0.8209008 
#> target_dist - cur_dist: 0.8109008 
#> target_dist - cur_dist: 0.8009008 
#> target_dist - cur_dist: 0.7909008 
#> target_dist - cur_dist: 0.7809008 
#> target_dist - cur_dist: 0.7709008 
#> target_dist - cur_dist: 0.7609008 
#> target_dist - cur_dist: 0.7509008 
#> target_dist - cur_dist: 0.7409008 
#> target_dist - cur_dist: 0.7309008 
#> target_dist - cur_dist: 0.7209008 
#> target_dist - cur_dist: 0.7109008 
#> target_dist - cur_dist: 0.7009008 
#> target_dist - cur_dist: 0.6909008 
#> target_dist - cur_dist: 0.6809008 
#> target_dist - cur_dist: 0.6709008 
#> target_dist - cur_dist: 0.6609008 
#> target_dist - cur_dist: 0.6509008 
#> target_dist - cur_dist: 0.6409008 
#> target_dist - cur_dist: 0.6309008 
#> target_dist - cur_dist: 0.6209008 
#> target_dist - cur_dist: 0.6109008 
#> target_dist - cur_dist: 0.6009008 
#> target_dist - cur_dist: 0.5909008 
#> target_dist - cur_dist: 0.5809008 
#> target_dist - cur_dist: 0.5709008 
#> target_dist - cur_dist: 0.5609008 
#> target_dist - cur_dist: 0.5509008 
#> target_dist - cur_dist: 0.5409008 
#> target_dist - cur_dist: 0.5309008 
#> target_dist - cur_dist: 0.5209008 
#> target_dist - cur_dist: 0.5109008 
#> target_dist - cur_dist: 0.5009008 
#> target_dist - cur_dist: 0.4909008 
#> target_dist - cur_dist: 0.4809008 
#> target_dist - cur_dist: 0.4709008 
#> target_dist - cur_dist: 0.4609008 
#> target_dist - cur_dist: 0.4509008 
#> target_dist - cur_dist: 0.4409008 
#> target_dist - cur_dist: 0.4309008 
#> target_dist - cur_dist: 0.4209008 
#> target_dist - cur_dist: 0.4109008 
#> target_dist - cur_dist: 0.4009008 
#> target_dist - cur_dist: 0.3909008 
#> target_dist - cur_dist: 0.3809008 
#> target_dist - cur_dist: 0.3709008 
#> target_dist - cur_dist: 0.3609008 
#> target_dist - cur_dist: 0.3509008 
#> target_dist - cur_dist: 0.3409008 
#> target_dist - cur_dist: 0.3309008 
#> target_dist - cur_dist: 0.3209008 
#> target_dist - cur_dist: 0.3109008 
#> target_dist - cur_dist: 0.3009008 
#> target_dist - cur_dist: 0.2909008 
#> target_dist - cur_dist: 0.2809008 
#> target_dist - cur_dist: 0.2709008 
#> target_dist - cur_dist: 0.2609008 
#> target_dist - cur_dist: 0.2509008 
#> target_dist - cur_dist: 0.2409008 
#> target_dist - cur_dist: 0.2309008 
#> target_dist - cur_dist: 0.2209008 
#> target_dist - cur_dist: 0.2109008 
#> target_dist - cur_dist: 0.2009008 
#> target_dist - cur_dist: 0.1909008 
#> target_dist - cur_dist: 0.1809008 
#> target_dist - cur_dist: 0.1709008 
#> target_dist - cur_dist: 0.1609008 
#> target_dist - cur_dist: 0.1509008 
#> target_dist - cur_dist: 0.1409008 
#> target_dist - cur_dist: 0.1309008 
#> target_dist - cur_dist: 0.1209008 
#> target_dist - cur_dist: 0.1109008 
#> target_dist - cur_dist: 0.1009008 
#> target_dist - cur_dist: 0.09090076 
#> target_dist - cur_dist: 0.08090076 
#> target_dist - cur_dist: 0.07090076 
#> target_dist - cur_dist: 0.06090076 
#> target_dist - cur_dist: 0.05090076 
#> target_dist - cur_dist: 0.04090076 
#> target_dist - cur_dist: 0.03090076 
#> target_dist - cur_dist: 0.02090076 
#> target_dist - cur_dist: 0.01090076 
#> target_dist - cur_dist: 0.0009007588 
#> generation:  dist =   1.392163 
#> target_dist - cur_dist: 1.392163 
#> target_dist - cur_dist: 1.382163 
#> target_dist - cur_dist: 1.372163 
#> target_dist - cur_dist: 1.362163 
#> target_dist - cur_dist: 1.352163 
#> target_dist - cur_dist: 1.342163 
#> target_dist - cur_dist: 1.332163 
#> target_dist - cur_dist: 1.322163 
#> target_dist - cur_dist: 1.312163 
#> target_dist - cur_dist: 1.302163 
#> target_dist - cur_dist: 1.292163 
#> target_dist - cur_dist: 1.282163 
#> target_dist - cur_dist: 1.272163 
#> target_dist - cur_dist: 1.262163 
#> target_dist - cur_dist: 1.252163 
#> target_dist - cur_dist: 1.242163 
#> target_dist - cur_dist: 1.232163 
#> target_dist - cur_dist: 1.222163 
#> target_dist - cur_dist: 1.212163 
#> target_dist - cur_dist: 1.202163 
#> target_dist - cur_dist: 1.192163 
#> target_dist - cur_dist: 1.182163 
#> target_dist - cur_dist: 1.172163 
#> target_dist - cur_dist: 1.162163 
#> target_dist - cur_dist: 1.152163 
#> target_dist - cur_dist: 1.142163 
#> target_dist - cur_dist: 1.132163 
#> target_dist - cur_dist: 1.122163 
#> target_dist - cur_dist: 1.112163 
#> target_dist - cur_dist: 1.102163 
#> target_dist - cur_dist: 1.092163 
#> target_dist - cur_dist: 1.082163 
#> target_dist - cur_dist: 1.072163 
#> target_dist - cur_dist: 1.062163 
#> target_dist - cur_dist: 1.052163 
#> target_dist - cur_dist: 1.042163 
#> target_dist - cur_dist: 1.032163 
#> target_dist - cur_dist: 1.022163 
#> target_dist - cur_dist: 1.012163 
#> target_dist - cur_dist: 1.002163 
#> target_dist - cur_dist: 0.9921634 
#> target_dist - cur_dist: 0.9821634 
#> target_dist - cur_dist: 0.9721634 
#> target_dist - cur_dist: 0.9621634 
#> target_dist - cur_dist: 0.9521634 
#> target_dist - cur_dist: 0.9421634 
#> target_dist - cur_dist: 0.9321634 
#> target_dist - cur_dist: 0.9221634 
#> target_dist - cur_dist: 0.9121634 
#> target_dist - cur_dist: 0.9021634 
#> target_dist - cur_dist: 0.8921634 
#> target_dist - cur_dist: 0.8821634 
#> target_dist - cur_dist: 0.8721634 
#> target_dist - cur_dist: 0.8621634 
#> target_dist - cur_dist: 0.8521634 
#> target_dist - cur_dist: 0.8421634 
#> target_dist - cur_dist: 0.8321634 
#> target_dist - cur_dist: 0.8221634 
#> target_dist - cur_dist: 0.8121634 
#> target_dist - cur_dist: 0.8021634 
#> target_dist - cur_dist: 0.7921634 
#> target_dist - cur_dist: 0.7821634 
#> target_dist - cur_dist: 0.7721634 
#> target_dist - cur_dist: 0.7621634 
#> target_dist - cur_dist: 0.7521634 
#> target_dist - cur_dist: 0.7421634 
#> target_dist - cur_dist: 0.7321634 
#> target_dist - cur_dist: 0.7221634 
#> target_dist - cur_dist: 0.7121634 
#> target_dist - cur_dist: 0.7021634 
#> target_dist - cur_dist: 0.6921634 
#> target_dist - cur_dist: 0.6821634 
#> target_dist - cur_dist: 0.6721634 
#> target_dist - cur_dist: 0.6621634 
#> target_dist - cur_dist: 0.6521634 
#> target_dist - cur_dist: 0.6421634 
#> target_dist - cur_dist: 0.6321634 
#> target_dist - cur_dist: 0.6221634 
#> target_dist - cur_dist: 0.6121634 
#> target_dist - cur_dist: 0.6021634 
#> target_dist - cur_dist: 0.5921634 
#> target_dist - cur_dist: 0.5821634 
#> target_dist - cur_dist: 0.5721634 
#> target_dist - cur_dist: 0.5621634 
#> target_dist - cur_dist: 0.5521634 
#> target_dist - cur_dist: 0.5421634 
#> target_dist - cur_dist: 0.5321634 
#> target_dist - cur_dist: 0.5221634 
#> target_dist - cur_dist: 0.5121634 
#> target_dist - cur_dist: 0.5021634 
#> target_dist - cur_dist: 0.4921634 
#> target_dist - cur_dist: 0.4821634 
#> target_dist - cur_dist: 0.4721634 
#> target_dist - cur_dist: 0.4621634 
#> target_dist - cur_dist: 0.4521634 
#> target_dist - cur_dist: 0.4421634 
#> target_dist - cur_dist: 0.4321634 
#> target_dist - cur_dist: 0.4221634 
#> target_dist - cur_dist: 0.4121634 
#> target_dist - cur_dist: 0.4021634 
#> target_dist - cur_dist: 0.3921634 
#> target_dist - cur_dist: 0.3821634 
#> target_dist - cur_dist: 0.3721634 
#> target_dist - cur_dist: 0.3621634 
#> target_dist - cur_dist: 0.3521634 
#> target_dist - cur_dist: 0.3421634 
#> target_dist - cur_dist: 0.3321634 
#> target_dist - cur_dist: 0.3221634 
#> target_dist - cur_dist: 0.3121634 
#> target_dist - cur_dist: 0.3021634 
#> target_dist - cur_dist: 0.2921634 
#> target_dist - cur_dist: 0.2821634 
#> target_dist - cur_dist: 0.2721634 
#> target_dist - cur_dist: 0.2621634 
#> target_dist - cur_dist: 0.2521634 
#> target_dist - cur_dist: 0.2421634 
#> target_dist - cur_dist: 0.2321634 
#> target_dist - cur_dist: 0.2221634 
#> target_dist - cur_dist: 0.2121634 
#> target_dist - cur_dist: 0.2021634 
#> target_dist - cur_dist: 0.1921634 
#> target_dist - cur_dist: 0.1821634 
#> target_dist - cur_dist: 0.1721634 
#> target_dist - cur_dist: 0.1621634 
#> target_dist - cur_dist: 0.1521634 
#> target_dist - cur_dist: 0.1421634 
#> target_dist - cur_dist: 0.1321634 
#> target_dist - cur_dist: 0.1221634 
#> target_dist - cur_dist: 0.1121634 
#> target_dist - cur_dist: 0.1021634 
#> target_dist - cur_dist: 0.09216336 
#> target_dist - cur_dist: 0.08216336 
#> target_dist - cur_dist: 0.07216336 
#> target_dist - cur_dist: 0.06216336 
#> target_dist - cur_dist: 0.05216336 
#> target_dist - cur_dist: 0.04216336 
#> target_dist - cur_dist: 0.03216336 
#> target_dist - cur_dist: 0.02216336 
#> target_dist - cur_dist: 0.01216336 
#> target_dist - cur_dist: 0.002163363 
#> generation:  dist =   1.3389 
#> target_dist - cur_dist: 1.3389 
#> target_dist - cur_dist: 1.3289 
#> target_dist - cur_dist: 1.3189 
#> target_dist - cur_dist: 1.3089 
#> target_dist - cur_dist: 1.2989 
#> target_dist - cur_dist: 1.2889 
#> target_dist - cur_dist: 1.2789 
#> target_dist - cur_dist: 1.2689 
#> target_dist - cur_dist: 1.2589 
#> target_dist - cur_dist: 1.2489 
#> target_dist - cur_dist: 1.2389 
#> target_dist - cur_dist: 1.2289 
#> target_dist - cur_dist: 1.2189 
#> target_dist - cur_dist: 1.2089 
#> target_dist - cur_dist: 1.1989 
#> target_dist - cur_dist: 1.1889 
#> target_dist - cur_dist: 1.1789 
#> target_dist - cur_dist: 1.1689 
#> target_dist - cur_dist: 1.1589 
#> target_dist - cur_dist: 1.1489 
#> target_dist - cur_dist: 1.1389 
#> target_dist - cur_dist: 1.1289 
#> target_dist - cur_dist: 1.1189 
#> target_dist - cur_dist: 1.1089 
#> target_dist - cur_dist: 1.0989 
#> target_dist - cur_dist: 1.0889 
#> target_dist - cur_dist: 1.0789 
#> target_dist - cur_dist: 1.0689 
#> target_dist - cur_dist: 1.0589 
#> target_dist - cur_dist: 1.0489 
#> target_dist - cur_dist: 1.0389 
#> target_dist - cur_dist: 1.0289 
#> target_dist - cur_dist: 1.0189 
#> target_dist - cur_dist: 1.0089 
#> target_dist - cur_dist: 0.9989 
#> target_dist - cur_dist: 0.9889 
#> target_dist - cur_dist: 0.9789 
#> target_dist - cur_dist: 0.9689 
#> target_dist - cur_dist: 0.9589 
#> target_dist - cur_dist: 0.9489 
#> target_dist - cur_dist: 0.9389 
#> target_dist - cur_dist: 0.9289 
#> target_dist - cur_dist: 0.9189 
#> target_dist - cur_dist: 0.9089 
#> target_dist - cur_dist: 0.8989 
#> target_dist - cur_dist: 0.8889 
#> target_dist - cur_dist: 0.8789 
#> target_dist - cur_dist: 0.8689 
#> target_dist - cur_dist: 0.8589 
#> target_dist - cur_dist: 0.8489 
#> target_dist - cur_dist: 0.8389 
#> target_dist - cur_dist: 0.8289 
#> target_dist - cur_dist: 0.8189 
#> target_dist - cur_dist: 0.8089 
#> target_dist - cur_dist: 0.7989 
#> target_dist - cur_dist: 0.7889 
#> target_dist - cur_dist: 0.7789 
#> target_dist - cur_dist: 0.7689 
#> target_dist - cur_dist: 0.7589 
#> target_dist - cur_dist: 0.7489 
#> target_dist - cur_dist: 0.7389 
#> target_dist - cur_dist: 0.7289 
#> target_dist - cur_dist: 0.7189 
#> target_dist - cur_dist: 0.7089 
#> target_dist - cur_dist: 0.6989 
#> target_dist - cur_dist: 0.6889 
#> target_dist - cur_dist: 0.6789 
#> target_dist - cur_dist: 0.6689 
#> target_dist - cur_dist: 0.6589 
#> target_dist - cur_dist: 0.6489 
#> target_dist - cur_dist: 0.6389 
#> target_dist - cur_dist: 0.6289 
#> target_dist - cur_dist: 0.6189 
#> target_dist - cur_dist: 0.6089 
#> target_dist - cur_dist: 0.5989 
#> target_dist - cur_dist: 0.5889 
#> target_dist - cur_dist: 0.5789 
#> target_dist - cur_dist: 0.5689 
#> target_dist - cur_dist: 0.5589 
#> target_dist - cur_dist: 0.5489 
#> target_dist - cur_dist: 0.5389 
#> target_dist - cur_dist: 0.5289 
#> target_dist - cur_dist: 0.5189 
#> target_dist - cur_dist: 0.5089 
#> target_dist - cur_dist: 0.4989 
#> target_dist - cur_dist: 0.4889 
#> target_dist - cur_dist: 0.4789 
#> target_dist - cur_dist: 0.4689 
#> target_dist - cur_dist: 0.4589 
#> target_dist - cur_dist: 0.4489 
#> target_dist - cur_dist: 0.4389 
#> target_dist - cur_dist: 0.4289 
#> target_dist - cur_dist: 0.4189 
#> target_dist - cur_dist: 0.4089 
#> target_dist - cur_dist: 0.3989 
#> target_dist - cur_dist: 0.3889 
#> target_dist - cur_dist: 0.3789 
#> target_dist - cur_dist: 0.3689 
#> target_dist - cur_dist: 0.3589 
#> target_dist - cur_dist: 0.3489 
#> target_dist - cur_dist: 0.3389 
#> target_dist - cur_dist: 0.3289 
#> target_dist - cur_dist: 0.3189 
#> target_dist - cur_dist: 0.3089 
#> target_dist - cur_dist: 0.2989 
#> target_dist - cur_dist: 0.2889 
#> target_dist - cur_dist: 0.2789 
#> target_dist - cur_dist: 0.2689 
#> target_dist - cur_dist: 0.2589 
#> target_dist - cur_dist: 0.2489 
#> target_dist - cur_dist: 0.2389 
#> target_dist - cur_dist: 0.2289 
#> target_dist - cur_dist: 0.2189 
#> target_dist - cur_dist: 0.2089 
#> target_dist - cur_dist: 0.1989 
#> target_dist - cur_dist: 0.1889 
#> target_dist - cur_dist: 0.1789 
#> target_dist - cur_dist: 0.1689 
#> target_dist - cur_dist: 0.1589 
#> target_dist - cur_dist: 0.1489 
#> target_dist - cur_dist: 0.1389 
#> target_dist - cur_dist: 0.1289 
#> target_dist - cur_dist: 0.1189 
#> target_dist - cur_dist: 0.1089 
#> target_dist - cur_dist: 0.09890005 
#> target_dist - cur_dist: 0.08890005 
#> target_dist - cur_dist: 0.07890005 
#> target_dist - cur_dist: 0.06890005 
#> target_dist - cur_dist: 0.05890005 
#> target_dist - cur_dist: 0.04890005 
#> target_dist - cur_dist: 0.03890005 
#> target_dist - cur_dist: 0.02890005 
#> target_dist - cur_dist: 0.01890005 
#> target_dist - cur_dist: 0.008900047 

#> [1]    6    1 1154

dim(interpolate(t1, 0.05))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.374828 
#> target_dist - cur_dist: 1.374828 
#> target_dist - cur_dist: 1.324828 
#> target_dist - cur_dist: 1.274828 
#> target_dist - cur_dist: 1.224828 
#> target_dist - cur_dist: 1.174828 
#> target_dist - cur_dist: 1.124828 
#> target_dist - cur_dist: 1.074828 
#> target_dist - cur_dist: 1.024828 
#> target_dist - cur_dist: 0.9748283 
#> target_dist - cur_dist: 0.9248283 
#> target_dist - cur_dist: 0.8748283 
#> target_dist - cur_dist: 0.8248283 
#> target_dist - cur_dist: 0.7748283 
#> target_dist - cur_dist: 0.7248283 
#> target_dist - cur_dist: 0.6748283 
#> target_dist - cur_dist: 0.6248283 
#> target_dist - cur_dist: 0.5748283 
#> target_dist - cur_dist: 0.5248283 
#> target_dist - cur_dist: 0.4748283 
#> target_dist - cur_dist: 0.4248283 
#> target_dist - cur_dist: 0.3748283 
#> target_dist - cur_dist: 0.3248283 
#> target_dist - cur_dist: 0.2748283 
#> target_dist - cur_dist: 0.2248283 
#> target_dist - cur_dist: 0.1748283 
#> target_dist - cur_dist: 0.1248283 
#> target_dist - cur_dist: 0.07482834 
#> target_dist - cur_dist: 0.02482834 
#> generation:  dist =   1.218371 
#> target_dist - cur_dist: 1.218371 
#> target_dist - cur_dist: 1.168371 
#> target_dist - cur_dist: 1.118371 
#> target_dist - cur_dist: 1.068371 
#> target_dist - cur_dist: 1.018371 
#> target_dist - cur_dist: 0.9683711 
#> target_dist - cur_dist: 0.9183711 
#> target_dist - cur_dist: 0.8683711 
#> target_dist - cur_dist: 0.8183711 
#> target_dist - cur_dist: 0.7683711 
#> target_dist - cur_dist: 0.7183711 
#> target_dist - cur_dist: 0.6683711 
#> target_dist - cur_dist: 0.6183711 
#> target_dist - cur_dist: 0.5683711 
#> target_dist - cur_dist: 0.5183711 
#> target_dist - cur_dist: 0.4683711 
#> target_dist - cur_dist: 0.4183711 
#> target_dist - cur_dist: 0.3683711 
#> target_dist - cur_dist: 0.3183711 
#> target_dist - cur_dist: 0.2683711 
#> target_dist - cur_dist: 0.2183711 
#> target_dist - cur_dist: 0.1683711 
#> target_dist - cur_dist: 0.1183711 
#> target_dist - cur_dist: 0.06837115 
#> target_dist - cur_dist: 0.01837115 
#> generation:  dist =   1.393842 
#> target_dist - cur_dist: 1.393842 
#> target_dist - cur_dist: 1.343842 
#> target_dist - cur_dist: 1.293842 
#> target_dist - cur_dist: 1.243842 
#> target_dist - cur_dist: 1.193842 
#> target_dist - cur_dist: 1.143842 
#> target_dist - cur_dist: 1.093842 
#> target_dist - cur_dist: 1.043842 
#> target_dist - cur_dist: 0.9938419 
#> target_dist - cur_dist: 0.9438419 
#> target_dist - cur_dist: 0.8938419 
#> target_dist - cur_dist: 0.8438419 
#> target_dist - cur_dist: 0.7938419 
#> target_dist - cur_dist: 0.7438419 
#> target_dist - cur_dist: 0.6938419 
#> target_dist - cur_dist: 0.6438419 
#> target_dist - cur_dist: 0.5938419 
#> target_dist - cur_dist: 0.5438419 
#> target_dist - cur_dist: 0.4938419 
#> target_dist - cur_dist: 0.4438419 
#> target_dist - cur_dist: 0.3938419 
#> target_dist - cur_dist: 0.3438419 
#> target_dist - cur_dist: 0.2938419 
#> target_dist - cur_dist: 0.2438419 
#> target_dist - cur_dist: 0.1938419 
#> target_dist - cur_dist: 0.1438419 
#> target_dist - cur_dist: 0.0938419 
#> target_dist - cur_dist: 0.0438419 
#> generation:  dist =   1.151928 
#> target_dist - cur_dist: 1.151928 
#> target_dist - cur_dist: 1.101928 
#> target_dist - cur_dist: 1.051928 
#> target_dist - cur_dist: 1.001928 
#> target_dist - cur_dist: 0.9519276 
#> target_dist - cur_dist: 0.9019276 
#> target_dist - cur_dist: 0.8519276 
#> target_dist - cur_dist: 0.8019276 
#> target_dist - cur_dist: 0.7519276 
#> target_dist - cur_dist: 0.7019276 
#> target_dist - cur_dist: 0.6519276 
#> target_dist - cur_dist: 0.6019276 
#> target_dist - cur_dist: 0.5519276 
#> target_dist - cur_dist: 0.5019276 
#> target_dist - cur_dist: 0.4519276 
#> target_dist - cur_dist: 0.4019276 
#> target_dist - cur_dist: 0.3519276 
#> target_dist - cur_dist: 0.3019276 
#> target_dist - cur_dist: 0.2519276 
#> target_dist - cur_dist: 0.2019276 
#> target_dist - cur_dist: 0.1519276 
#> target_dist - cur_dist: 0.1019276 
#> target_dist - cur_dist: 0.05192763 
#> target_dist - cur_dist: 0.001927631 
#> generation:  dist =   1.144572 
#> target_dist - cur_dist: 1.144572 
#> target_dist - cur_dist: 1.094572 
#> target_dist - cur_dist: 1.044572 
#> target_dist - cur_dist: 0.994572 
#> target_dist - cur_dist: 0.944572 
#> target_dist - cur_dist: 0.894572 
#> target_dist - cur_dist: 0.844572 
#> target_dist - cur_dist: 0.794572 
#> target_dist - cur_dist: 0.744572 
#> target_dist - cur_dist: 0.694572 
#> target_dist - cur_dist: 0.644572 
#> target_dist - cur_dist: 0.594572 
#> target_dist - cur_dist: 0.544572 
#> target_dist - cur_dist: 0.494572 
#> target_dist - cur_dist: 0.444572 
#> target_dist - cur_dist: 0.394572 
#> target_dist - cur_dist: 0.344572 
#> target_dist - cur_dist: 0.294572 
#> target_dist - cur_dist: 0.244572 
#> target_dist - cur_dist: 0.194572 
#> target_dist - cur_dist: 0.144572 
#> target_dist - cur_dist: 0.09457201 
#> target_dist - cur_dist: 0.04457201 
#> generation:  dist =   1.372434 
#> target_dist - cur_dist: 1.372434 
#> target_dist - cur_dist: 1.322434 
#> target_dist - cur_dist: 1.272434 
#> target_dist - cur_dist: 1.222434 
#> target_dist - cur_dist: 1.172434 
#> target_dist - cur_dist: 1.122434 
#> target_dist - cur_dist: 1.072434 
#> target_dist - cur_dist: 1.022434 
#> target_dist - cur_dist: 0.9724345 
#> target_dist - cur_dist: 0.9224345 
#> target_dist - cur_dist: 0.8724345 
#> target_dist - cur_dist: 0.8224345 
#> target_dist - cur_dist: 0.7724345 
#> target_dist - cur_dist: 0.7224345 
#> target_dist - cur_dist: 0.6724345 
#> target_dist - cur_dist: 0.6224345 
#> target_dist - cur_dist: 0.5724345 
#> target_dist - cur_dist: 0.5224345 
#> target_dist - cur_dist: 0.4724345 
#> target_dist - cur_dist: 0.4224345 
#> target_dist - cur_dist: 0.3724345 
#> target_dist - cur_dist: 0.3224345 
#> target_dist - cur_dist: 0.2724345 
#> target_dist - cur_dist: 0.2224345 
#> target_dist - cur_dist: 0.1724345 
#> target_dist - cur_dist: 0.1224345 
#> target_dist - cur_dist: 0.07243449 
#> target_dist - cur_dist: 0.02243449 
#> generation:  dist =   1.090901 
#> target_dist - cur_dist: 1.090901 
#> target_dist - cur_dist: 1.040901 
#> target_dist - cur_dist: 0.9909008 
#> target_dist - cur_dist: 0.9409008 
#> target_dist - cur_dist: 0.8909008 
#> target_dist - cur_dist: 0.8409008 
#> target_dist - cur_dist: 0.7909008 
#> target_dist - cur_dist: 0.7409008 
#> target_dist - cur_dist: 0.6909008 
#> target_dist - cur_dist: 0.6409008 
#> target_dist - cur_dist: 0.5909008 
#> target_dist - cur_dist: 0.5409008 
#> target_dist - cur_dist: 0.4909008 
#> target_dist - cur_dist: 0.4409008 
#> target_dist - cur_dist: 0.3909008 
#> target_dist - cur_dist: 0.3409008 
#> target_dist - cur_dist: 0.2909008 
#> target_dist - cur_dist: 0.2409008 
#> target_dist - cur_dist: 0.1909008 
#> target_dist - cur_dist: 0.1409008 
#> target_dist - cur_dist: 0.09090076 
#> target_dist - cur_dist: 0.04090076 
#> generation:  dist =   1.392163 
#> target_dist - cur_dist: 1.392163 
#> target_dist - cur_dist: 1.342163 
#> target_dist - cur_dist: 1.292163 
#> target_dist - cur_dist: 1.242163 
#> target_dist - cur_dist: 1.192163 
#> target_dist - cur_dist: 1.142163 
#> target_dist - cur_dist: 1.092163 
#> target_dist - cur_dist: 1.042163 
#> target_dist - cur_dist: 0.9921634 
#> target_dist - cur_dist: 0.9421634 
#> target_dist - cur_dist: 0.8921634 
#> target_dist - cur_dist: 0.8421634 
#> target_dist - cur_dist: 0.7921634 
#> target_dist - cur_dist: 0.7421634 
#> target_dist - cur_dist: 0.6921634 
#> target_dist - cur_dist: 0.6421634 
#> target_dist - cur_dist: 0.5921634 
#> target_dist - cur_dist: 0.5421634 
#> target_dist - cur_dist: 0.4921634 
#> target_dist - cur_dist: 0.4421634 
#> target_dist - cur_dist: 0.3921634 
#> target_dist - cur_dist: 0.3421634 
#> target_dist - cur_dist: 0.2921634 
#> target_dist - cur_dist: 0.2421634 
#> target_dist - cur_dist: 0.1921634 
#> target_dist - cur_dist: 0.1421634 
#> target_dist - cur_dist: 0.09216336 
#> target_dist - cur_dist: 0.04216336 
#> generation:  dist =   1.3389 
#> target_dist - cur_dist: 1.3389 
#> target_dist - cur_dist: 1.2889 
#> target_dist - cur_dist: 1.2389 
#> target_dist - cur_dist: 1.1889 
#> target_dist - cur_dist: 1.1389 
#> target_dist - cur_dist: 1.0889 
#> target_dist - cur_dist: 1.0389 
#> target_dist - cur_dist: 0.9889 
#> target_dist - cur_dist: 0.9389 
#> target_dist - cur_dist: 0.8889 
#> target_dist - cur_dist: 0.8389 
#> target_dist - cur_dist: 0.7889 
#> target_dist - cur_dist: 0.7389 
#> target_dist - cur_dist: 0.6889 
#> target_dist - cur_dist: 0.6389 
#> target_dist - cur_dist: 0.5889 
#> target_dist - cur_dist: 0.5389 
#> target_dist - cur_dist: 0.4889 
#> target_dist - cur_dist: 0.4389 
#> target_dist - cur_dist: 0.3889 
#> target_dist - cur_dist: 0.3389 
#> target_dist - cur_dist: 0.2889 
#> target_dist - cur_dist: 0.2389 
#> target_dist - cur_dist: 0.1889 
#> target_dist - cur_dist: 0.1389 
#> target_dist - cur_dist: 0.08890005 
#> target_dist - cur_dist: 0.03890005 

#> [1]   6   1 234

dim(interpolate(t1, 0.1))

#> target_dist - cur_dist: 0 
#> generation:  dist =   1.374828 
#> target_dist - cur_dist: 1.374828 
#> target_dist - cur_dist: 1.274828 
#> target_dist - cur_dist: 1.174828 
#> target_dist - cur_dist: 1.074828 
#> target_dist - cur_dist: 0.9748283 
#> target_dist - cur_dist: 0.8748283 
#> target_dist - cur_dist: 0.7748283 
#> target_dist - cur_dist: 0.6748283 
#> target_dist - cur_dist: 0.5748283 
#> target_dist - cur_dist: 0.4748283 
#> target_dist - cur_dist: 0.3748283 
#> target_dist - cur_dist: 0.2748283 
#> target_dist - cur_dist: 0.1748283 
#> target_dist - cur_dist: 0.07482834 
#> generation:  dist =   1.218371 
#> target_dist - cur_dist: 1.218371 
#> target_dist - cur_dist: 1.118371 
#> target_dist - cur_dist: 1.018371 
#> target_dist - cur_dist: 0.9183711 
#> target_dist - cur_dist: 0.8183711 
#> target_dist - cur_dist: 0.7183711 
#> target_dist - cur_dist: 0.6183711 
#> target_dist - cur_dist: 0.5183711 
#> target_dist - cur_dist: 0.4183711 
#> target_dist - cur_dist: 0.3183711 
#> target_dist - cur_dist: 0.2183711 
#> target_dist - cur_dist: 0.1183711 
#> target_dist - cur_dist: 0.01837115 
#> generation:  dist =   1.393842 
#> target_dist - cur_dist: 1.393842 
#> target_dist - cur_dist: 1.293842 
#> target_dist - cur_dist: 1.193842 
#> target_dist - cur_dist: 1.093842 
#> target_dist - cur_dist: 0.9938419 
#> target_dist - cur_dist: 0.8938419 
#> target_dist - cur_dist: 0.7938419 
#> target_dist - cur_dist: 0.6938419 
#> target_dist - cur_dist: 0.5938419 
#> target_dist - cur_dist: 0.4938419 
#> target_dist - cur_dist: 0.3938419 
#> target_dist - cur_dist: 0.2938419 
#> target_dist - cur_dist: 0.1938419 
#> target_dist - cur_dist: 0.0938419 
#> generation:  dist =   1.151928 
#> target_dist - cur_dist: 1.151928 
#> target_dist - cur_dist: 1.051928 
#> target_dist - cur_dist: 0.9519276 
#> target_dist - cur_dist: 0.8519276 
#> target_dist - cur_dist: 0.7519276 
#> target_dist - cur_dist: 0.6519276 
#> target_dist - cur_dist: 0.5519276 
#> target_dist - cur_dist: 0.4519276 
#> target_dist - cur_dist: 0.3519276 
#> target_dist - cur_dist: 0.2519276 
#> target_dist - cur_dist: 0.1519276 
#> target_dist - cur_dist: 0.05192763 
#> generation:  dist =   1.144572 
#> target_dist - cur_dist: 1.144572 
#> target_dist - cur_dist: 1.044572 
#> target_dist - cur_dist: 0.944572 
#> target_dist - cur_dist: 0.844572 
#> target_dist - cur_dist: 0.744572 
#> target_dist - cur_dist: 0.644572 
#> target_dist - cur_dist: 0.544572 
#> target_dist - cur_dist: 0.444572 
#> target_dist - cur_dist: 0.344572 
#> target_dist - cur_dist: 0.244572 
#> target_dist - cur_dist: 0.144572 
#> target_dist - cur_dist: 0.04457201 
#> generation:  dist =   1.372434 
#> target_dist - cur_dist: 1.372434 
#> target_dist - cur_dist: 1.272434 
#> target_dist - cur_dist: 1.172434 
#> target_dist - cur_dist: 1.072434 
#> target_dist - cur_dist: 0.9724345 
#> target_dist - cur_dist: 0.8724345 
#> target_dist - cur_dist: 0.7724345 
#> target_dist - cur_dist: 0.6724345 
#> target_dist - cur_dist: 0.5724345 
#> target_dist - cur_dist: 0.4724345 
#> target_dist - cur_dist: 0.3724345 
#> target_dist - cur_dist: 0.2724345 
#> target_dist - cur_dist: 0.1724345 
#> target_dist - cur_dist: 0.07243449 
#> generation:  dist =   1.090901 
#> target_dist - cur_dist: 1.090901 
#> target_dist - cur_dist: 0.9909008 
#> target_dist - cur_dist: 0.8909008 
#> target_dist - cur_dist: 0.7909008 
#> target_dist - cur_dist: 0.6909008 
#> target_dist - cur_dist: 0.5909008 
#> target_dist - cur_dist: 0.4909008 
#> target_dist - cur_dist: 0.3909008 
#> target_dist - cur_dist: 0.2909008 
#> target_dist - cur_dist: 0.1909008 
#> target_dist - cur_dist: 0.09090076 
#> generation:  dist =   1.392163 
#> target_dist - cur_dist: 1.392163 
#> target_dist - cur_dist: 1.292163 
#> target_dist - cur_dist: 1.192163 
#> target_dist - cur_dist: 1.092163 
#> target_dist - cur_dist: 0.9921634 
#> target_dist - cur_dist: 0.8921634 
#> target_dist - cur_dist: 0.7921634 
#> target_dist - cur_dist: 0.6921634 
#> target_dist - cur_dist: 0.5921634 
#> target_dist - cur_dist: 0.4921634 
#> target_dist - cur_dist: 0.3921634 
#> target_dist - cur_dist: 0.2921634 
#> target_dist - cur_dist: 0.1921634 
#> target_dist - cur_dist: 0.09216336 
#> generation:  dist =   1.3389 
#> target_dist - cur_dist: 1.3389 
#> target_dist - cur_dist: 1.2389 
#> target_dist - cur_dist: 1.1389 
#> target_dist - cur_dist: 1.0389 
#> target_dist - cur_dist: 0.9389 
#> target_dist - cur_dist: 0.8389 
#> target_dist - cur_dist: 0.7389 
#> target_dist - cur_dist: 0.6389 
#> target_dist - cur_dist: 0.5389 
#> target_dist - cur_dist: 0.4389 
#> target_dist - cur_dist: 0.3389 
#> target_dist - cur_dist: 0.2389 
#> target_dist - cur_dist: 0.1389 
#> target_dist - cur_dist: 0.03890005 

#> [1]   6   1 119